AreaInter {spatstat.model}	R Documentation

The Area Interaction Point Process Model

Description

Creates an instance of the Area Interaction point process model (Widom-Rowlinson penetrable spheres model) which can then be fitted to point pattern data.

Usage

  AreaInter(r)

Arguments

r
The radius of the discs in the area interaction process

Details

This function defines the interpoint interaction structure of a point process called the Widom-Rowlinson penetrable sphere model or area-interaction process. It can be used to fit this model to point pattern data.

The function ppm(), which fits point process models to point pattern data, requires an argument of class "interact" describing the interpoint interaction structure of the model to be fitted. The appropriate description of the area interaction structure is yielded by the function AreaInter(). See the examples below.

In standard form, the area-interaction process (Widom and Rowlinson, 1970; Baddeley and Van Lieshout, 1995) with disc radius r, intensity parameter κ and interaction parameter γ is a point process with probability density

f(x[1],…,x[n]) = α . κ^n(x) . γ^(-A(x))

for a point pattern x, where x[1],…,x[n] represent the points of the pattern, n(x) is the number of points in the pattern, and A(x) is the area of the region formed by the union of discs of radius r centred at the points x[1],…,x[n]. Here α is a normalising constant.

The interaction parameter γ can be any positive number. If γ = 1 then the model reduces to a Poisson process with intensity κ. If γ < 1 then the process is regular, while if γ > 1 the process is clustered. Thus, an area interaction process can be used to model either clustered or regular point patterns. Two points interact if the distance between them is less than 2 * r.

The standard form of the model, shown above, is a little complicated to interpret in practical applications. For example, each isolated point of the pattern x contributes a factor κ * γ^(-π * r^2) to the probability density.

In spatstat, the model is parametrised in a different form, which is easier to interpret. In canonical scale-free form, the probability density is rewritten as

f(x_1,…,x_n) = α . β^n(x) η^(-C(x))

where β is the new intensity parameter, η is the new interaction parameter, and C(x) = B(x) - n(x) is the interaction potential. Here

B(x) = A(x)/(π * r^2)

is the normalised area (so that the discs have unit area). In this formulation, each isolated point of the pattern contributes a factor β to the probability density (so the first order trend is β). The quantity C(x) is a true interaction potential, in the sense that C(x) = 0 if the point pattern x does not contain any points that lie close together (closer than 2*r units apart).

When a new point u is added to an existing point pattern x, the rescaled potential -C(x) increases by a value between 0 and 1. The increase is zero if u is not close to any point of x. The increase is 1 if the disc of radius r centred at u is completely contained in the union of discs of radius r centred at the data points x[i]. Thus, the increase in potential is a measure of how close the new point u is to the existing pattern x. Addition of the point u contributes a factor β * η^δ to the probability density, where δ is the increase in potential.

The old parameters κ,γ of the standard form are related to the new parameters β,η of the canonical scale-free form, by

β = κ * γ^(-π * r^2)= κ / η

and

η = γ^(π * r^2)

provided γ and κ are positive and finite.

In the canonical scale-free form, the parameter η can take any nonnegative value. The value η = 1 again corresponds to a Poisson process, with intensity β. If η < 1 then the process is regular, while if η > 1 the process is clustered. The value η = 0 corresponds to a hard core process with hard core radius r (interaction distance 2r).

The nonstationary area interaction process is similar except that the contribution of each individual point x[i] is a function β(x[i]) of location, rather than a constant beta.

Note the only argument of AreaInter() is the disc radius r. When r is fixed, the model becomes an exponential family. The canonical parameters log(β) and log(η) are estimated by ppm(), not fixed in AreaInter().

Value

An object of class "interact" describing the interpoint interaction structure of the area-interaction process with disc radius r.

Warnings

The interaction distance of this process is equal to 2 * r. Two discs of radius r overlap if their centres are closer than 2 * r units apart.

The estimate of the interaction parameter η is unreliable if the interaction radius r is too small or too large. In these situations the model is approximately Poisson so that η is unidentifiable. As a rule of thumb, one can inspect the empty space function of the data, computed by Fest. The value F(r) of the empty space function at the interaction radius r should be between 0.2 and 0.8.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

References

Baddeley, A.J. and Van Lieshout, M.N.M. (1995). Area-interaction point processes. Annals of the Institute of Statistical Mathematics 47 (1995) 601–619.

Widom, B. and Rowlinson, J.S. (1970). New model for the study of liquid-vapor phase transitions. The Journal of Chemical Physics 52 (1970) 1670–1684.

See Also

ppm, pairwise.family, ppm.object

ragsAreaInter and rmh for simulation of area-interaction models.

Examples

   

   # prints a sensible description of itself
   AreaInter(r=0.1)

   # Note the reach is twice the radius
   reach(AreaInter(r=1))

   # Fit the stationary area interaction process to Swedish Pines data
   ppm(swedishpines ~1, AreaInter(r=7))

   # Fit the stationary area interaction process to `cells'
   ppm(cells ~1, AreaInter(r=0.06))
   # eta=0 indicates hard core process.

   # Fit a nonstationary area interaction with log-cubic polynomial trend
   
     ppm(swedishpines ~polynom(x/10,y/10,3), AreaInter(r=7))
   

   

[Package spatstat.model version 3.3-4 Index]