Concom {spatstat.model}	R Documentation

The Connected Component Process Model

Description

Creates an instance of the Connected Component point process model which can then be fitted to point pattern data.

Usage

  Concom(r)

Arguments

r
Threshold distance

Details

This function defines the interpoint interaction structure of a point process called the connected component process. It can be used to fit this model to point pattern data.

The function ppm(), which fits point process models to point pattern data, requires an argument of class "interact" describing the interpoint interaction structure of the model to be fitted. The appropriate description of the connected component interaction is yielded by the function Concom(). See the examples below.

In standard form, the connected component process (Baddeley and Moller, 1989) with disc radius r, intensity parameter κ and interaction parameter γ is a point process with probability density

f(x[1],…,x[n]) = α . κ^n(x) . γ^(-C(x))

for a point pattern x, where x[1],…,x[n] represent the points of the pattern, n(x) is the number of points in the pattern, and C(x) is defined below. Here α is a normalising constant.

To define the term C(x), suppose that we construct a planar graph by drawing an edge between each pair of points x[i],x[j] which are less than r units apart. Two points belong to the same connected component of this graph if they are joined by a path in the graph. Then C(x) is the number of connected components of the graph.

The interaction parameter γ can be any positive number. If γ = 1 then the model reduces to a Poisson process with intensity κ. If γ < 1 then the process is regular, while if γ > 1 the process is clustered. Thus, a connected-component interaction process can be used to model either clustered or regular point patterns.

In spatstat, the model is parametrised in a different form, which is easier to interpret. In canonical form, the probability density is rewritten as

f(x_1,…,x_n) = α . β^n(x) γ^(-U(x))

where β is the new intensity parameter and U(x) = C(x) - n(x) is the interaction potential. In this formulation, each isolated point of the pattern contributes a factor β to the probability density (so the first order trend is β). The quantity U(x) is a true interaction potential, in the sense that U(x) = 0 if the point pattern x does not contain any points that lie close together.

When a new point u is added to an existing point pattern x, the rescaled potential -U(x) increases by zero or a positive integer. The increase is zero if u is not close to any point of x. The increase is a positive integer k if there are k different connected components of x that lie close to u. Addition of the point u contributes a factor β * η^δ to the probability density, where δ is the increase in potential.

If desired, the original parameter κ can be recovered from the canonical parameter by κ = β * γ.

The nonstationary connected component process is similar except that the contribution of each individual point x[i] is a function β(x[i]) of location, rather than a constant beta.

Note the only argument of Concom() is the threshold distance r. When r is fixed, the model becomes an exponential family. The canonical parameters log(β) and log(γ) are estimated by ppm(), not fixed in Concom().

Value

An object of class "interact" describing the interpoint interaction structure of the connected component process with disc radius r.

Edge correction

The interaction distance of this process is infinite. There are no well-established procedures for edge correction for fitting such models, and accordingly the model-fitting function ppm will give an error message saying that the user must specify an edge correction. A reasonable solution is to use the border correction at the same distance r, as shown in the Examples.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

References

Baddeley, A.J. and Moller, J. (1989) Nearest-neighbour Markov point processes and random sets. International Statistical Review 57, 89–121.

See Also

ppm, pairwise.family, ppm.object

Examples

   # prints a sensible description of itself
   Concom(r=0.1)

   # Fit the stationary connected component process to redwood data
   ppm(redwood ~1, Concom(r=0.07), rbord=0.07)

   # Fit the stationary connected component process to `cells' data
   ppm(cells ~1, Concom(r=0.06), rbord=0.06)
   # eta=0 indicates hard core process.

   # Fit a nonstationary connected component model
   # with log-cubic polynomial trend
   
     ppm(swedishpines ~polynom(x/10,y/10,3), Concom(r=7), rbord=7)
   

[Package spatstat.model version 3.3-4 Index]