DiggleGratton {spatstat.model}	R Documentation

Diggle-Gratton model

Description

Creates an instance of the Diggle-Gratton pairwise interaction point process model, which can then be fitted to point pattern data.

Usage

  DiggleGratton(delta=NA, rho)

Arguments

delta
lower threshold δ

rho
upper threshold ρ

Details

Diggle and Gratton (1984, pages 208-210) introduced the pairwise interaction point process with pair potential h(t) of the form

h(t) = ((t - δ)/(ρ - δ))^κ, { } δ ≤ t ≤ ρ

with h(t) = 0 for t < δ and h(t) = 1 for t > ρ. Here δ, ρ and κ are parameters.

Note that we use the symbol κ where Diggle and Gratton (1984) and Diggle, Gates and Stibbard (1987) use β, since in spatstat we reserve the symbol β for an intensity parameter.

The parameters must all be nonnegative, and must satisfy δ ≤ ρ.

The potential is inhibitory, i.e.\ this model is only appropriate for regular point patterns. The strength of inhibition increases with κ. For κ=0 the model is a hard core process with hard core radius δ. For κ=Inf the model is a hard core process with hard core radius ρ.

The irregular parameters δ, ρ must be given in the call to DiggleGratton, while the regular parameter κ will be estimated.

If the lower threshold delta is missing or NA, it will be estimated from the data when ppm is called. The estimated value of delta is the minimum nearest neighbour distance multiplied by n/(n+1), where n is the number of data points.

Value

An object of class "interact" describing the interpoint interaction structure of a point process.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

References

Diggle, P.J., Gates, D.J. and Stibbard, A. (1987) A nonparametric estimator for pairwise-interaction point processes. Biometrika 74, 763 – 770.

Diggle, P.J. and Gratton, R.J. (1984) Monte Carlo methods of inference for implicit statistical models. Journal of the Royal Statistical Society, series B 46, 193 – 212.

See Also

ppm, ppm.object, Pairwise

Examples

   ppm(cells ~1, DiggleGratton(0.05, 0.1))

[Package spatstat.model version 3.3-4 Index]