Extract.influence.ppm {spatstat.model}	R Documentation

Extract Subset of Influence Object

Description

Extract a subset of an influence object, or extract the influence values at specified locations.

Usage

  ## S3 method for class 'influence.ppm'
x[i, ...]

Arguments

x
A influence object (of class "influence.ppm") computed by influence.ppm.

i
Subset index (passed to [.ppp). Either a spatial window (object of class "owin") or an integer index.

...
Ignored.

Details

An object of class "influence.ppm" contains the values of the likelihood influence for a point process model, computed by influence.ppm. This is effectively a marked point pattern obtained by marking each of the original data points with its likelihood influence.

This function extracts a designated subset of the influence values, either as another influence object, or as a vector of numeric values.

The function [.influence.ppm is a method for [ for the class "influence.ppm". The argument i should be an index applicable to a point pattern. It may be either a spatial window (object of class "owin") or a sequence index. The result will be another influence object (of class influence.ppm).

To extract the influence values as a numeric vector, use marks(as.ppp(x)).

Value

Another object of class "influence.ppm".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

influence.ppm.

Examples

   fit <- ppm(cells, ~x)
   infl <- influence(fit)
   b <- owin(c(0.1, 0.3), c(0.2, 0.4))
   infl[b]
   infl[1:5]
   marks(as.ppp(infl))[1:3]

[Package spatstat.model version 3.3-4 Index]