Extract.msr {spatstat.model}	R Documentation

Extract Subset of Signed or Vector Measure

Description

Extract a subset of a signed measure or vector-valued measure.

Usage

## S3 method for class 'msr'
x[i, j, ...]

Arguments

x
A signed or vector measure. An object of class "msr" (see msr).

i
Object defining the subregion or subset to be extracted. Either a spatial window (an object of class "owin"), or a pixel image with logical values, or any type of index that applies to a matrix.

j
Subset index selecting the vector coordinates to be extracted, if x is a vector-valued measure.

...
Ignored.

Details

This operator extracts a subset of the data which determines the signed measure or vector-valued measure x. The result is another measure.

Value

An object of class "msr".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

msr

Examples

   X <- rpoispp(function(x,y) { exp(3+3*x) })
   fit <- ppm(X ~x+y)
   rp <- residuals(fit, type="pearson")
   rs <- residuals(fit, type="score")

   rp[square(0.5)]
   rs[ , 2:3]

[Package spatstat.model version 3.3-4 Index]