HierStraussHard {spatstat.model}	R Documentation

The Hierarchical Strauss Hard Core Point Process Model

Description

Creates an instance of the hierarchical Strauss-hard core point process model which can then be fitted to point pattern data.

Usage

  HierStraussHard(iradii, hradii=NULL, types=NULL, archy=NULL)

Arguments

iradii
Matrix of interaction radii

hradii
Optional matrix of hard core distances

types
Optional; vector of all possible types (i.e. the possible levels of the marks variable in the data)

archy
Optional: the hierarchical order. See Details.

Details

This is a hierarchical point process model for a multitype point pattern (Hogmander and Sarkka, 1999; Grabarnik and Sarkka, 2009). It is appropriate for analysing multitype point pattern data in which the types are ordered so that the points of type j depend on the points of type 1,2,...,j-1.

The hierarchical version of the (stationary) Strauss hard core process with m types, with interaction radii r[i,j], hard core distances h[i,j] and parameters beta[j] and gamma[i,j] is a point process in which each point of type j contributes a factor beta[j] to the probability density of the point pattern, and a pair of points of types i and j closer than r[i,j] units apart contributes a factor gamma[i,j] to the density provided i <= j. If any pair of points of types i and j lies closer than h[i,j] units apart, the configuration of points is impossible (probability density zero).

The nonstationary hierarchical Strauss hard core process is similar except that the contribution of each individual point x[i] is a function beta(x[i]) of location and type, rather than a constant beta.

The function ppm(), which fits point process models to point pattern data, requires an argument of class "interact" describing the interpoint interaction structure of the model to be fitted. The appropriate description of the hierarchical Strauss hard core process pairwise interaction is yielded by the function HierStraussHard(). See the examples below.

The argument types need not be specified in normal use. It will be determined automatically from the point pattern data set to which the HierStraussHard interaction is applied, when the user calls ppm. However, the user should be confident that the ordering of types in the dataset corresponds to the ordering of rows and columns in the matrix radii.

The argument archy can be used to specify a hierarchical ordering of the types. It can be either a vector of integers or a character vector matching the possible types. The default is the sequence 1,2, ..., m meaning that type j depends on types 1,2, ..., j-1.

The matrices iradii and hradii must be square, with entries which are either positive numbers or zero or NA. A value of zero or NA indicates that no interaction term should be included for this combination of types.

Note that only the interaction radii and hard core distances are specified in HierStraussHard. The canonical parameters log(beta[j]) and log(gamma[i,j]) are estimated by ppm(), not fixed in HierStraussHard().

Value

An object of class "interact" describing the interpoint interaction structure of the hierarchical Strauss-hard core process with interaction radii iradii[i,j] and hard core distances hradii[i,j].

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

, Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk.

References

Grabarnik, P. and Sarkka, A. (2009) Modelling the spatial structure of forest stands by multivariate point processes with hierarchical interactions. Ecological Modelling 220, 1232–1240.

Hogmander, H. and Sarkka, A. (1999) Multitype spatial point patterns with hierarchical interactions. Biometrics 55, 1051–1058.

See Also

MultiStraussHard for the corresponding symmetrical interaction.

HierHard, HierStrauss.

Examples

   r <- matrix(c(30, NA, 40, 30), nrow=2,ncol=2)
   h <- matrix(c(4, NA, 10, 15), 2, 2)
   HierStraussHard(r, h)
   # prints a sensible description of itself
   ppm(ants ~1, HierStraussHard(r, h))
   # fit the stationary hierarchical Strauss-hard core process to ants data

[Package spatstat.model version 3.3-4 Index]