Hybrid {spatstat.model}	R Documentation

Hybrid Interaction Point Process Model

Description

Creates an instance of a hybrid point process model which can then be fitted to point pattern data.

Usage

Hybrid(...)

Arguments

...
Two or more interactions (objects of class "interact") or objects which can be converted to interactions. See Details.

Details

A hybrid (Baddeley, Turner, Mateu and Bevan, 2013) is a point process model created by combining two or more point process models, or an interpoint interaction created by combining two or more interpoint interactions.

The hybrid of two point processes, with probability densities f(x) and g(x) respectively, is the point process with probability density

h(x) = c * f(x) * g(x)

where c is a normalising constant.

Equivalently, the hybrid of two point processes with conditional intensities lambda(u,x) and kappa(u,x) is the point process with conditional intensity

phi(u,x) = lambda(u,x) * kappa(u,x).

The hybrid of m > 3 point processes is defined in a similar way.

The function ppm, which fits point process models to point pattern data, requires an argument of class "interact" describing the interpoint interaction structure of the model to be fitted. The appropriate description of a hybrid interaction is yielded by the function Hybrid().

The arguments ... will be interpreted as interpoint interactions (objects of class "interact") and the result will be the hybrid of these interactions. Each argument must either be an interpoint interaction (object of class "interact"), or a point process model (object of class "ppm") from which the interpoint interaction will be extracted.

The arguments ... may also be given in the form name=value. This is purely cosmetic: it can be used to attach simple mnemonic names to the component interactions, and makes the printed output from print.ppm neater.

Value

An object of class "interact" describing an interpoint interaction structure.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

References

Baddeley, A., Turner, R., Mateu, J. and Bevan, A. (2013) Hybrids of Gibbs point process models and their implementation. Journal of Statistical Software 55:11, 1–43. DOI: 10.18637/jss.v055.i11

See Also

ppm

Examples

  Hybrid(Strauss(0.1), Geyer(0.2, 3))

  Hybrid(Ha=Hardcore(0.05), St=Strauss(0.1), Ge=Geyer(0.2, 3))

  fit <- ppm(redwood, ~1, Hybrid(A=Strauss(0.02), B=Geyer(0.1, 2)))
  fit

  ctr <- rmhcontrol(nrep=5e4, expand=1)
  plot(simulate(fit, control=ctr))

  # hybrid components can be models (including hybrid models)
  Hybrid(fit, S=Softcore(0.5))

  # plot.fii only works if every component is a pairwise interaction
  fit2 <- ppm(swedishpines, ~1, Hybrid(DG=DiggleGratton(2,10), S=Strauss(5)))
  plot(fitin(fit2))
  plot(fitin(fit2), separate=TRUE, mar.panel=rep(4,4))

[Package spatstat.model version 3.3-4 Index]