Kmodel.dppm {spatstat.model}	R Documentation

K-function or Pair Correlation Function of a Determinantal Point Process Model

Description

Returns the theoretical K-function or theoretical pair correlation function of a determinantal point process model as a function of one argument r.

Usage

   ## S3 method for class 'dppm'
Kmodel(model, ...)

   ## S3 method for class 'dppm'
pcfmodel(model, ...)

   ## S3 method for class 'detpointprocfamily'
Kmodel(model, ...)

   ## S3 method for class 'detpointprocfamily'
pcfmodel(model, ...)

Arguments

model
Model of class "detpointprocfamily" or "dppm".

...
Ignored (not quite true – there is some undocumented internal use)

Value

A function in the R language, with one numeric argument r, that can be used to evaluate the theoretical K-function or pair correlation function of the model at distances r.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

Examples

  model <- dppMatern(lambda=100, alpha=.01, nu=1, d=2)
  KMatern <- Kmodel(model)
  pcfMatern <- pcfmodel(model)
  plot(KMatern, xlim = c(0,0.05))
  plot(pcfMatern, xlim = c(0,0.05))

[Package spatstat.model version 3.3-4 Index]