Kmodel.kppm {spatstat.model}	R Documentation

K Function or Pair Correlation Function of Cluster Model or Cox model

Description

Returns the theoretical K function or the pair correlation function of a cluster point process model or Cox point process model.

Usage

   ## S3 method for class 'kppm'
Kmodel(model, ...)

   ## S3 method for class 'kppm'
pcfmodel(model, ...)

Arguments

model
A fitted cluster point process model (object of class "kppm") typically obtained from the model-fitting algorithm kppm.

...
Ignored.

Details

For certain types of point process models, it is possible to write down a mathematical expression for the K function or the pair correlation function of the model. In particular this is possible for a fitted cluster point process model (object of class "kppm" obtained from kppm).

The functions Kmodel and pcfmodel are generic. The functions documented here are the methods for the class "kppm".

The return value is a function in the R language, which takes one argument r. Evaluation of this function, on a numeric vector r, yields values of the desired K function or pair correlation function at these distance values.

Value

A function in the R language, which takes one argument r.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

Kest or pcf to estimate the K function or pair correlation function nonparametrically from data.

kppm to fit cluster models.

Kmodel for the generic functions.

Kmodel.ppm for the method for Gibbs processes.

Examples

  fit <- kppm(redwood, ~x, "MatClust")
  K <- Kmodel(fit)
  K(c(0.1, 0.2))
  curve(K(x), from=0, to=0.25)

[Package spatstat.model version 3.3-4 Index]