LambertW {spatstat.model}	R Documentation

Lambert's W Function

Description

Computes Lambert's W-function.

Usage

LambertW(x)

Arguments

x
Vector of nonnegative numbers.

Details

Lambert's W-function is the inverse function of f(y) = y * exp(y). That is, W is the function such that

W(x) * exp(W(x)) = x

This command LambertW computes W(x) for each entry in the argument x. If the library gsl has been installed, then the function lambert_W0 in that library is invoked. Otherwise, values of the W-function are computed by root-finding, using the function uniroot.

Computation using gsl is about 100 times faster.

If any entries of x are infinite or NA, the corresponding results are NA.

Value

Numeric vector.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

References

Corless, R, Gonnet, G, Hare, D, Jeffrey, D and Knuth, D (1996), On the Lambert W function. Computational Mathematics, 5, 325–359.

Roy, R and Olver, F (2010), Lambert W function. In Olver, F, Lozier, D and Boisvert, R (eds.), NIST Handbook of Mathematical Functions, Cambridge University Press.

Examples

   LambertW(exp(1))

[Package spatstat.model version 3.3-4 Index]