Penttinen {spatstat.model}	R Documentation

Penttinen Interaction

Description

Creates an instance of the Penttinen pairwise interaction point process model, which can then be fitted to point pattern data.

Usage

  Penttinen(r)

Arguments

r
circle radius

Details

Penttinen (1984, Example 2.1, page 18), citing Cormack (1979), described the pairwise interaction point process with interaction factor

h(d) = exp(theta * A(d)) = gamma^(A(d))

between each pair of points separated by a distance $d$. Here A(d) is the area of intersection between two discs of radius r separated by a distance d, normalised so that A(0) = 1.

The scale of interaction is controlled by the disc radius r: two points interact if they are closer than 2 * r apart. The strength of interaction is controlled by the canonical parameter theta, which must be less than or equal to zero, or equivalently by the parameter gamma = exp(theta), which must lie between 0 and 1.

The potential is inhibitory, i.e.\ this model is only appropriate for regular point patterns. For gamma=0 the model is a hard core process with hard core diameter 2 * r. For gamma=1 the model is a Poisson process.

The irregular parameter r must be given in the call to Penttinen, while the regular parameter theta will be estimated.

This model can be considered as a pairwise approximation to the area-interaction model AreaInter.

Value

An object of class "interact" describing the interpoint interaction structure of a point process.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

References

Cormack, R.M. (1979) Spatial aspects of competition between individuals. Pages 151–212 in Spatial and Temporal Analysis in Ecology, eds. R.M. Cormack and J.K. Ord, International Co-operative Publishing House, Fairland, MD, USA.

Penttinen, A. (1984) Modelling Interaction in Spatial Point Patterns: Parameter Estimation by the Maximum Likelihood Method. Jyvaskyla Studies in Computer Science, Economics and Statistics 7, University of Jyvaskyla, Finland.

See Also

ppm, ppm.object, Pairwise, AreaInter.

Examples

   fit <- ppm(cells ~ 1, Penttinen(0.07))
   fit
   reach(fit) # interaction range is circle DIAMETER

[Package spatstat.model version 3.3-4 Index]