Smooth.msr {spatstat.model}	R Documentation

Smooth a Signed or Vector-Valued Measure

Description

Apply kernel smoothing to a signed measure or vector-valued measure.

Usage

 ## S3 method for class 'msr'
Smooth(X, ..., drop=TRUE)

Arguments

X
Object of class "msr" representing a signed measure or vector-valued measure.

...
Arguments passed to density.ppp controlling the smoothing bandwidth and the pixel resolution.

drop
Logical. If TRUE (the default), the result of smoothing a scalar-valued measure is a pixel image. If FALSE, the result of smoothing a scalar-valued measure is a list containing one pixel image.

Details

This function applies kernel smoothing to a signed measure or vector-valued measure X. The Gaussian kernel is used.

The object X would typically have been created by residuals.ppm or msr.

Value

A pixel image or a list of pixel images. For scalar-valued measures, a pixel image (object of class "im") provided drop=TRUE. For vector-valued measures (or if drop=FALSE), a list of pixel images; the list also belongs to the class "solist" so that it can be printed and plotted.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

References

Baddeley, A., Turner, R., Moller, J. and Hazelton, M. (2005) Residual analysis for spatial point processes. Journal of the Royal Statistical Society, Series B 67, 617–666.

Baddeley, A., Moller, J. and Pakes, A.G. (2008) Properties of residuals for spatial point processes. Annals of the Institute of Statistical Mathematics 60, 627–649.

See Also

Smooth, msr, plot.msr

Examples

   X <- rpoispp(function(x,y) { exp(3+3*x) })
   fit <- ppm(X, ~x+y)
   rp <- residuals(fit, type="pearson")
   rs <- residuals(fit, type="score")

   plot(Smooth(rp))
   plot(Smooth(rs))

[Package spatstat.model version 3.3-4 Index]