Triplets {spatstat.model}	R Documentation

The Triplet Point Process Model

Description

Creates an instance of Geyer's triplet interaction point process model which can then be fitted to point pattern data.

Usage

  Triplets(r)

Arguments

r
The interaction radius of the Triplets process

Details

The (stationary) Geyer triplet process (Geyer, 1999) with interaction radius r and parameters beta and gamma is the point process in which each point contributes a factor beta to the probability density of the point pattern, and each triplet of close points contributes a factor gamma to the density. A triplet of close points is a group of 3 points, each pair of which is closer than r units apart.

Thus the probability density is

f(x_1,…,x_n) = alpha . beta^n(x) gamma^s(x)

where x[1],…,x[n] represent the points of the pattern, n(x) is the number of points in the pattern, s(x) is the number of unordered triples of points that are closer than r units apart, and alpha is the normalising constant.

The interaction parameter gamma must be less than or equal to 1 so that this model describes an “ordered” or “inhibitive” pattern.

The nonstationary Triplets process is similar except that the contribution of each individual point x[i] is a function beta(x[i]) of location, rather than a constant beta.

The function ppm(), which fits point process models to point pattern data, requires an argument of class "interact" describing the interpoint interaction structure of the model to be fitted. The appropriate description of the Triplets process pairwise interaction is yielded by the function Triplets(). See the examples below.

Note the only argument is the interaction radius r. When r is fixed, the model becomes an exponential family. The canonical parameters log(beta) and log(gamma) are estimated by ppm(), not fixed in Triplets().

Value

An object of class "interact" describing the interpoint interaction structure of the Triplets process with interaction radius r.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

References

Geyer, C.J. (1999) Likelihood Inference for Spatial Point Processes. Chapter 3 in O.E. Barndorff-Nielsen, W.S. Kendall and M.N.M. Van Lieshout (eds) Stochastic Geometry: Likelihood and Computation, Chapman and Hall / CRC, Monographs on Statistics and Applied Probability, number 80. Pages 79–140.

See Also

ppm, triplet.family, ppm.object

Examples

   Triplets(r=0.1)
   # prints a sensible description of itself

   ppm(cells ~1, Triplets(r=0.2))
   # fit the stationary Triplets process to `cells'

   
     ppm(cells ~polynom(x,y,3), Triplets(r=0.2))
     # fit a nonstationary Triplets process with log-cubic polynomial trend
   

[Package spatstat.model version 3.3-4 Index]