as.function.leverage.ppm {spatstat.model}	R Documentation

Convert Leverage Object to Function of Coordinates

Description

Converts an object of class "leverage.ppm" to a function of the x and y coordinates.

Usage

 ## S3 method for class 'leverage.ppm'
as.function(x, ...)

Arguments

x
Object of class "leverage.ppm" produced by leverage.ppm.

...
Ignored.

Details

An object of class "leverage.ppm" represents the leverage function of a fitted point process model. This command converts the object to a function(x,y) where the arguments x and y are (vectors of) spatial coordinates. This function returns the leverage values at the specified locations (calculated by referring to the nearest location where the leverage has been computed).

Value

A function in the R language, also belonging to the class "funxy".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

as.im.leverage.ppm

Examples

  X <- rpoispp(function(x,y) { exp(3+3*x) })
  fit <- ppm(X ~x+y)
  lev <- leverage(fit)
  f <- as.function(lev)
  
  f(0.2, 0.3)  # evaluate at (x,y) coordinates
  y <- f(X)    # evaluate at a point pattern

[Package spatstat.model version 3.3-4 Index]