clusterfield.kppm {spatstat.model}	R Documentation

Field of clusters

Description

Calculate the superposition of cluster kernels at the location of a point pattern.

Usage

  ## S3 method for class 'kppm'
clusterfield(model, locations = NULL, ...)

Arguments

model
Cluster model. Either a fitted cluster model (object of class "kppm"), a character string specifying the type of cluster model, or a function defining the cluster kernel. See Details.

locations
A point pattern giving the locations of the kernels. Defaults to the centroid of the observation window for the "kppm" method and to the center of a unit square otherwise.

...
Additional arguments passed to density.ppp or the cluster kernel. See Details.

Details

The function clusterfield is generic, with a method for "kppm" (described here) and methods for "character" and "function".

The method clusterfield.kppm extracts the relevant information from the fitted model and calls clusterfield.function.

The calculations are performed by density.ppp and ... arguments are passed thereto for control over the pixel resolution etc. (These arguments are then passed on to pixellate.ppp and as.mask.)

Value

A pixel image (object of class "im").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

kppm,

clusterfield.

Examples

  fit <- kppm(redwood~1, "Thomas")
  Z <- clusterfield(fit, eps = 0.01)

[Package spatstat.model version 3.3-4 Index]