clusterkernel.kppm {spatstat.model}	R Documentation

Extract Cluster Offspring Kernel

Description

Given a fitted cluster point process model, this command returns the probability density of the cluster offspring.

Usage

## S3 method for class 'kppm'
clusterkernel(model, ...)

Arguments

model
Cluster model. Either a fitted cluster or Cox model (object of class "kppm"), or a character string specifying the type of cluster model.

...
Parameter values for the model, when model is a character string.

Details

Given a cluster point process model, this command returns a function(x,y) giving the two-dimensional probability density of the cluster offspring points assuming a cluster parent located at the origin.

The function clusterkernel is generic, with methods for class "kppm" (described here) and "character" (described in clusterkernel.character).

Value

A function in the R language with arguments x,y,....

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

clusterkernel.character, clusterfield, kppm

Examples

  fit <- kppm(redwood ~ x, "MatClust")
  f <- clusterkernel(fit)
  f(0.05, 0.02)

[Package spatstat.model version 3.3-4 Index]