coef.slrm {spatstat.model}	R Documentation

Coefficients of Fitted Spatial Logistic Regression Model

Description

Extracts the coefficients (parameters) from a fitted Spatial Logistic Regression model.

Usage

  ## S3 method for class 'slrm'
coef(object, ...)

Arguments

object
a fitted spatial logistic regression model. An object of class "slrm".

...
Ignored.

Details

This is a method for coef for fitted spatial logistic regression models (objects of class "slrm", usually obtained from the function slrm).

It extracts the fitted canonical parameters, i.e.\ the coefficients in the linear predictor of the spatial logistic regression.

Value

Numeric vector of coefficients.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

slrm

Examples

  X <- rpoispp(42)
  fit <- slrm(X ~ x+y)
  coef(fit)

[Package spatstat.model version 3.3-4 Index]