dfbetas.ppm {spatstat.model}	R Documentation

Parameter Influence Measure

Description

Computes the deletion influence measure for each parameter in a fitted point process model.

Usage

## S3 method for class 'ppm'
dfbetas(model, ...,
       drop = FALSE, iScore=NULL, iHessian=NULL, iArgs=NULL)

Arguments

model
Fitted point process model (object of class "ppm").

...
Ignored, except for the arguments dimyx and eps which are passed to as.mask to control the spatial resolution of the image of the density component.

drop
Logical. Whether to include (drop=FALSE) or exclude (drop=TRUE) contributions from quadrature points that were not used to fit the model.

iScore,iHessian
Components of the score vector and Hessian matrix for the irregular parameters, if required. See Details.

iArgs
List of extra arguments for the functions iScore, iHessian if required.

Details

Given a fitted spatial point process model, this function computes the influence measure for each parameter, as described in Baddeley, Chang and Song (2013) and Baddeley, Rubak and Turner (2019).

This is a method for the generic function dfbetas.

The influence measure for each parameter θ is a signed measure in two-dimensional space. It consists of a discrete mass on each data point (i.e. each point in the point pattern to which the model was originally fitted) and a continuous density at all locations. The mass at a data point represents the change in the fitted value of the parameter θ that would occur if this data point were to be deleted. The density at other non-data locations represents the effect (on the fitted value of θ) of deleting these locations (and their associated covariate values) from the input to the fitting procedure.

If the point process model trend has irregular parameters that were fitted (using ippm) then the influence calculation requires the first and second derivatives of the log trend with respect to the irregular parameters. The argument iScore should be a list, with one entry for each irregular parameter, of R functions that compute the partial derivatives of the log trend (i.e. log intensity or log conditional intensity) with respect to each irregular parameter. The argument iHessian should be a list, with p^2 entries where p is the number of irregular parameters, of R functions that compute the second order partial derivatives of the log trend with respect to each pair of irregular parameters.

Value

An object of class "msr" representing a signed or vector-valued measure. This object can be printed and plotted.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

References

Baddeley, A. and Chang, Y.M. and Song, Y. (2013) Leverage and influence diagnostics for spatial point process models. Scandinavian Journal of Statistics 40, 86–104.

Baddeley, A., Rubak, E. and Turner, R. (2019) Leverage and influence diagnostics for Gibbs spatial point processes. Spatial Statistics 29, 15–48.

See Also

leverage.ppm, influence.ppm, ppmInfluence.

See msr for information on how to use a measure.

Examples

   
   X <- rpoispp(function(x,y) { exp(3+3*x) })
   fit <- ppm(X ~x+y)
   
   plot(dfbetas(fit))
   plot(Smooth(dfbetas(fit)))
   

[Package spatstat.model version 3.3-4 Index]