dffit.ppm {spatstat.model}	R Documentation

Case Deletion Effect Measure of Fitted Model

Description

Computes the case deletion effect measure DFFIT for a fitted model.

Usage

dffit(object, ...)

## S3 method for class 'ppm'
dffit(object, ..., collapse = FALSE, dfb = NULL)

Arguments

object
A fitted model, such as a point process model (object of class "ppm").

...
Additional arguments passed to dfbetas.ppm.

collapse
Logical value specifying whether to collapse the vector-valued measure to a scalar-valued measure by adding all the components.

dfb
Optional. The result of dfbetas(object), if it has already been computed.

Details

The case deletion effect measure DFFIT is a model diagnostic traditionally used for regression models. In that context, DFFIT[i,j] is the negative change, in the value of the jth term in the linear predictor, that would occur if the ith data value was deleted. It is closely related to the diagnostic DFBETA.

For a spatial point process model, dffit computes the analogous spatial case deletion diagnostic, described in Baddeley, Rubak and Turner (2019).

Value

A measure (object of class "msr").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

References

Baddeley, A., Rubak, E. and Turner, R. (2019) Leverage and influence diagnostics for Gibbs spatial point processes. Spatial Statistics 29, 15–48.

See Also

dfbetas.ppm

Examples

   
   X <- rpoispp(function(x,y) { exp(3+3*x) })
   fit <- ppm(X ~x+y)
   
   plot(dffit(fit))
   plot(dffit(fit, collapse=TRUE))
   
   

[Package spatstat.model version 3.3-4 Index]