dppMatern {spatstat.model}	R Documentation

Whittle-Matern Determinantal Point Process Model

Description

Function generating an instance of the Whittle-Matern determinantal point process model

Usage

dppMatern(...)

Arguments

...
arguments of the form tag=value specifying the parameters. See Details.

Details

The Whittle-Matern DPP is defined in (Lavancier, Moller and Rubak, 2015) The possible parameters are:

the intensity lambda as a positive numeric

the scale parameter alpha as a positive numeric

the shape parameter nu as a positive numeric (artificially required to be less than 20 in the code for numerical stability)

the dimension d as a positive integer

Value

An object of class "detpointprocfamily".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

References

Lavancier, F. Moller, J. and Rubak, E. (2015) Determinantal point process models and statistical inference Journal of the Royal Statistical Society, Series B 77, 853–977.

See Also

dppBessel, dppCauchy, dppGauss, dppPowerExp

Examples

m <- dppMatern(lambda=100, alpha=.02, nu=1, d=2)

[Package spatstat.model version 3.3-4 Index]