dppapproxkernel {spatstat.model}	R Documentation

Approximate Determinantal Point Process Kernel

Description

Returns an approximation to the kernel of a determinantal point process, as a function of one argument x.

Usage

dppapproxkernel(model, trunc = 0.99, W = NULL)

Arguments

model
Object of class "detpointprocfamily".

trunc
Numeric specifying how the model truncation is performed. See Details section of simulate.detpointprocfamily.

W
Optional window – undocumented at the moment.

Value

A function

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

[Package spatstat.model version 3.3-4 Index]