dppapproxpcf {spatstat.model}	R Documentation

Approximate Pair Correlation Function of Determinantal Point Process Model

Description

Returns an approximation to the theoretical pair correlation function of a determinantal point process model, as a function of one argument x.

Usage

dppapproxpcf(model, trunc = 0.99, W = NULL)

Arguments

model
Object of class "detpointprocfamily".

trunc
Numeric value specifying how the model truncation is performed. See Details section of simulate.detpointprocfamily.

W
Optional window – undocumented at the moment.

Details

This function is usually NOT needed for anything. It only exists for investigative purposes.

Value

A function in the R language, with one numeric argument x, that returns the value of the approximate pair correlation at distances x.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

Examples

f <- dppapproxpcf(dppMatern(lambda = 100, alpha=.028, nu=1, d=2))
plot(f, xlim = c(0,0.1))

[Package spatstat.model version 3.3-4 Index]