dppeigen {spatstat.model}	R Documentation

Internal function calculating eig and index

Description

This function is mainly for internal package use and is usually not called by the user.

Usage

dppeigen(model, trunc, Wscale, stationary = FALSE)

Arguments

model
object of class "detpointprocfamily"

trunc
numeric giving the truncation

Wscale
numeric giving the scale of the window relative to a unit box

stationary
logical indicating whether the stationarity of the model should be used (only works in dimension 2).

Value

A list

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

[Package spatstat.model version 3.3-4 Index]