dppkernel {spatstat.model}	R Documentation

Extract Kernel from Determinantal Point Process Model Object

Description

Returns the kernel of a determinantal point process model as a function of one argument x.

Usage

dppkernel(model, ...)

Arguments

model
Model of class "detpointprocfamily".

...
Arguments passed to dppapproxkernel if the exact kernel is unknown

Value

A function

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

Examples

kernelMatern <- dppkernel(dppMatern(lambda = 100, alpha=.01, nu=1, d=2))
plot(kernelMatern, xlim = c(0,0.1))

[Package spatstat.model version 3.3-4 Index]