dppparbounds {spatstat.model}	R Documentation

Parameter Bound for a Determinantal Point Process Model

Description

Returns the lower and upper bound for a specific parameter of a determinantal point process model when all other parameters are fixed.

Usage

dppparbounds(model, name, ...)

Arguments

model
Model of class "detpointprocfamily".

name
name of the parameter for which the bound should be computed.

...
Additional arguments passed to the parbounds function of the given model

Value

A data.frame containing lower and upper bounds.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

Examples

model <- dppMatern(lambda=100, alpha=.01, nu=1, d=2)
dppparbounds(model, "lambda")

[Package spatstat.model version 3.3-4 Index]