dppspecden {spatstat.model}	R Documentation

Extract Spectral Density from Determinantal Point Process Model Object

Description

Returns the spectral density of a determinantal point process model as a function of one argument x.

Usage

dppspecden(model)

Arguments

model
Model of class "detpointprocfamily".

Value

A function

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

See Also

dppspecdenrange

Examples

model <- dppMatern(lambda = 100, alpha=.01, nu=1, d=2)
dppspecden(model)

[Package spatstat.model version 3.3-4 Index]