emend {spatstat.model}	R Documentation

Force Model to be Valid

Description

Check whether a model is valid, and if not, find the nearest model which is valid.

Usage

emend(object, ...)

Arguments

object
A statistical model, belonging to some class.

...
Arguments passed to methods.

Details

The function emend is generic, and has methods for several classes of statistical models in the spatstat package (mostly point process models). Its purpose is to check whether a given model is valid (for example, that none of the model parameters are NA) and, if not, to find the nearest model which is valid.

See the methods for more information.

Value

Another model of the same kind.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

emend.ppm, valid.

[Package spatstat.model version 3.3-4 Index]