fitted.mppm {spatstat.model}	R Documentation

Fitted Conditional Intensity for Multiple Point Process Model

Description

Given a point process model fitted to multiple point patterns, compute the fitted conditional intensity of the model at the points of each data pattern, or at the points of the quadrature schemes used to fit the model.

Usage

 ## S3 method for class 'mppm'
fitted(object, ..., type = "lambda", dataonly = FALSE)

Arguments

object
The fitted model. An object of class "mppm" obtained from mppm.

...
Ignored.

type
Type of fitted values: either "trend" for the spatial trend, or "lambda" or "cif" for the conditional intensity.

dataonly
If TRUE, fitted values are computed only for the points of the data point patterns. If FALSE, fitted values are computed for the points of the quadrature schemes used to fit the model.

Details

This function evaluates the conditional intensity lambdahat(u,x) or spatial trend bhat(u) of the fitted point process model for certain locations u, for each of the original point patterns x to which the model was fitted.

The locations u at which the fitted conditional intensity/trend is evaluated, are the points of the quadrature schemes used to fit the model in mppm. They include the data points (the points of the original point pattern datasets) and other “dummy” points in the window of observation.

Use predict.mppm to compute the fitted conditional intensity at other locations or with other values of the explanatory variables.

Value

A list of vectors (one for each row of the original hyperframe, i.e. one vector for each of the original point patterns) containing the values of the fitted conditional intensity or (if type="trend") the fitted spatial trend.

Entries in these vector correspond to the quadrature points (data or dummy points) used to fit the model. The quadrature points can be extracted from object by quad.mppm(object).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Ida-Maria Sintorn and Leanne Bischoff. Implemented by Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

References

Baddeley, A., Rubak, E. and Turner, R. (2015) Spatial Point Patterns: Methodology and Applications with R. Chapman and Hall/CRC Press.

See Also

mppm, predict.mppm

Examples

   model <- mppm(Bugs ~ x, data=hyperframe(Bugs=waterstriders),
                           interaction=Strauss(7))
   cifs <- fitted(model)

[Package spatstat.model version 3.3-4 Index]