formula.ppm {spatstat.model}	R Documentation

Model Formulae for Gibbs Point Process Models

Description

Extract the trend formula, or the terms in the trend formula, in a fitted Gibbs point process model.

Usage

## S3 method for class 'ppm'
formula(x, ...)
## S3 method for class 'ppm'
terms(x, ...)

Arguments

x
An object of class "ppm", representing a fitted point process model.

...
Arguments passed to other methods.

Details

These functions are methods for the generic commands formula and terms for the class "ppm".

An object of class "ppm" represents a fitted Poisson or Gibbs point process model. It is obtained from the model-fitting function ppm.

The method formula.ppm extracts the trend formula from the fitted model x (the formula originally specified as the argument trend to ppm). The method terms.ppm extracts the individual terms in the trend formula.

Value

See the help files for the corresponding generic functions.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

ppm, as.owin, coef.ppm, extractAIC.ppm, fitted.ppm, logLik.ppm, model.frame.ppm, model.matrix.ppm, plot.ppm, predict.ppm, residuals.ppm, simulate.ppm, summary.ppm, update.ppm, vcov.ppm.

Examples

  fit <- ppm(cells, ~x)
  formula(fit)
  terms(fit)  

[Package spatstat.model version 3.3-4 Index]