hardcoredist {spatstat.model}	R Documentation

Extract the Hard Core Distance of a Point Process Model

Description

Extract or compute the hard core distance of a point process model.

Usage

hardcoredist(x, ...)

## S3 method for class 'fii'
hardcoredist(x, ..., epsilon = 0)

## S3 method for class 'ppm'
hardcoredist(x, ..., epsilon = 0)

Arguments

x
An object representing a point process model (class "ppm") or the interaction structure of a point process (class "fii") or similar.

...
Additional arguments passed to methods.

epsilon
Tolerance for defining the hard core.

Details

A point process model has a hard core distance h if it is impossible for two random points to lie closer than the distance h apart.

The function hardcoredist is generic, with methods for objects of class "ppm" (point process models) and "fii" (fitted point process interactions). It extracts or computes the hard core distance.

If epsilon is specified, then the code calculates the largest distance at which the interaction factor is smaller than epsilon, implying that points are unlikely to occur closer than this distance.

The result is zero if the model does not have a hard core distance.

Value

A single numeric value, or for multitype point processes, a numeric matrix giving the hard core distances for each pair of types of points.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

Examples

  m <- ppm(cells~1, Hardcore())
  hardcoredist(m)

[Package spatstat.model version 3.3-4 Index]