harmonise.msr {spatstat.model}	R Documentation

Make Measures Compatible

Description

Convert several measures to a common quadrature scheme

Usage

## S3 method for class 'msr'
harmonise(...)

Arguments

...
Any number of measures (objects of class "msr").

Details

This function makes any number of measures compatible, by converting them all to a common quadrature scheme.

The command harmonise is generic. This is the method for objects of class "msr".

Value

A list, of length equal to the number of arguments ..., whose entries are measures.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

harmonise, msr

Examples

  fit1 <- ppm(cells ~ x)
  fit2 <- ppm(rpoispp(ex=cells) ~ x)
  m1 <- residuals(fit1)
  m2 <- residuals(fit2)
  harmonise(m1, m2)
  s1 <- residuals(fit1, type="score")
  s2 <- residuals(fit2, type="score")
  harmonise(s1, s2)

[Package spatstat.model version 3.3-4 Index]