hybrid.family {spatstat.model}	R Documentation

Hybrid Interaction Family

Description

An object describing the family of all hybrid interactions.

Details

Advanced Use Only!

This structure would not normally be touched by the user. It describes the family of all hybrid point process models.

If you need to create a specific hybrid interaction model for use in modelling, use the function Hybrid.

Value

Object of class "isf", see isf.object.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

See Also

Use Hybrid to make hybrid interactions.

Other families: pairwise.family, pairsat.family, ord.family, inforder.family.

[Package spatstat.model version 3.3-4 Index]