ic.kppm {spatstat.model}	R Documentation

Model selection criteria for the intensity function of a point process

Description

Information criteria for selecting the intensity function model of a Poisson, cluster or Cox point process.

Usage

  ic(object)

  ## S3 method for class 'ppm'
ic(object)

  ## S3 method for class 'kppm'
ic(object)

Arguments

object
Fitted point process model (object of class "ppm" or "kppm").

Details

This function returns information criteria for selecting the intensity function model of a Poisson, Cox or cluster point process fitted by first order composite likelihood (i.e. using the Poisson likelihood function).

Degrees of freedom df for the information criteria are given by the trace of S^{-1} Σ where S is the sensitivity matrix and Σ is the variance matrix for the log composite likelihood score function. In case of a Poisson process, df is the number of parameters in the model for the intensity function.

The composite Bayesian information criterion (cbic) is -2 * ll + log(n) * df where ll is the maximal log first-order composite likelihood (Poisson loglikelihood for the intensity function) and n is the observed number of points. It reduces to the BIC criterion in case of a Poisson process.

The composite information criterion (cic) is - 2 * ll + 2 * df and reduces to the AIC in case of a Poisson process.

NOTE: the information criteria are for selecting the intensity function model (a set of covariates) within a given model class. They cannot be used to choose among different types of cluster or Cox point process models (e.g. can not be used to choose between Thomas and LGCP models).

Value

A list with entries loglike, cbic, cic and df. Here loglike is the fitted log first-order composite likelihood, cbic is composite Bayesian information criterion, cic is is the composite likelihood criterion and df is the adjusted degrees of freedom for the fitted intensity function model.

Author(s)

Achmad Choiruddin, Jean-Francois Coeurjolly and Rasmus Waagepetersen.

References

Choiruddin, A., Coeurjolly, J.F. and Waagepetersen, R. (2020) Information criteria for inhomogeneous spatial point processes. Australian and New Zealand Journal of Statistics. To appear.

See Also

kppm

Examples

  if(interactive()) {

    # model with one covariate
    fit1 <- kppm(bei~elev,data=bei.extra)
    ic1  <- ic(fit1)

    # model with two covariates
    fit2 <- kppm(bei~elev+grad,data=bei.extra)
    ic2  <- ic(fit2)
  
    # smallest cbic for fit1 but smallest cic for fit2
  }

[Package spatstat.model version 3.3-4 Index]