improve.kppm {spatstat.model}	R Documentation

Improve Intensity Estimate of Fitted Cluster Point Process Model

Description

Update the fitted intensity of a fitted cluster point process model.

Usage

improve.kppm(object, type=c("quasi", "wclik1", "clik1"), rmax = NULL,
             eps.rmax = 0.01, dimyx = 50, maxIter = 100, tolerance = 1e-06,
             fast = TRUE, vcov = FALSE, fast.vcov = FALSE, verbose = FALSE,
                          save.internals = FALSE)

Arguments

object
Fitted cluster point process model (object of class "kppm").

type
A character string indicating the method of estimation. Current options are "clik1", "wclik1" and "quasi" for, respectively, first order composite (Poisson) likelihood, weighted first order composite likelihood and quasi-likelihood.

rmax
Optional. The dependence range. Not usually specified by the user.

eps.rmax
Numeric. A small positive number which is used to determine rmax from the tail behaviour of the pair correlation function. Namely rmax is the smallest value of r at which (g(r)-1)/(g(0)-1) falls below eps.rmax. Ignored if rmax is provided.

dimyx
Pixel array dimensions. See Details.

maxIter
Integer. Maximum number of iterations of iterative weighted least squares (Fisher scoring).

tolerance
Numeric. Tolerance value specifying when to stop iterative weighted least squares (Fisher scoring).

fast
Logical value indicating whether tapering should be used to make the computations faster (requires the package Matrix).

vcov
Logical value indicating whether to calculate the asymptotic variance covariance/matrix.

fast.vcov
Logical value indicating whether tapering should be used for the variance/covariance matrix to make the computations faster (requires the package Matrix). Caution: This is expected to underestimate the true asymptotic variances/covariances.

verbose
A logical indicating whether the details of computations should be printed.

save.internals
A logical indicating whether internal quantities should be saved in the returned object (mostly for development purposes).

Details

This function reestimates the intensity parameters in a fitted "kppm" object. If type="clik1" estimates are based on the first order composite (Poisson) likelihood, which ignores dependence between the points. Note that type="clik1" is mainly included for testing purposes and is not recommended for the typical user; instead the more efficient kppm with improve.type="none" should be used.

When type="quasi" or type="wclik1" the dependence structure between the points is incorporated in the estimation procedure by using the estimated pair correlation function in the estimating equation.

In all cases the estimating equation is based on dividing the observation window into small subregions and count the number of points in each subregion. To do this the observation window is first converted into a digital mask by as.mask where the resolution is controlled by the argument dimyx. The computational time grows with the cube of the number of subregions, so fine grids may take very long to compute (or even run out of memory).

Value

A fitted cluster point process model of class "kppm".

Author(s)

Abdollah Jalilian jalilian@razi.ac.ir and Rasmus Plenge Waagepetersen rw@math.auc.dk. Adapted for spatstat by Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Ege Rubak rubak@math.aau.dk.

References

Waagepetersen, R. (2007) An estimating function approach to inference for inhomogeneous Neyman-Scott processes, Biometrics, 63, 252-258.

Guan, Y. and Shen, Y. (2010) A weighted estimating equation approach to inference for inhomogeneous spatial point processes, Biometrika, 97, 867-880.

Guan, Y., Jalilian, A. and Waagepetersen, R. (2015) Quasi-likelihood for spatial point processes. Journal of the Royal Statistical Society, Series B 77, 677–697.

See Also

ppm, kppm, improve.kppm

Examples

  # fit a Thomas process using minimum contrast estimation method 
  # to model interaction between points of the pattern
  fit0 <- kppm(bei ~ elev + grad, data = bei.extra)

  # fit the log-linear intensity model with quasi-likelihood method
  fit1 <- improve.kppm(fit0, type="quasi")

  # compare
  coef(fit0)
  coef(fit1)

[Package spatstat.model version 3.3-4 Index]