influence.ppm {spatstat.model}	R Documentation

Influence Measure for Spatial Point Process Model

Description

Computes the influence measure for a fitted spatial point process model.

Usage

## S3 method for class 'ppm'
influence(model, ...,
        drop = FALSE, iScore=NULL, iHessian=NULL, iArgs=NULL)

Arguments

model
Fitted point process model (object of class "ppm").

...
Ignored.

drop
Logical. Whether to include (drop=FALSE) or exclude (drop=TRUE) contributions from quadrature points that were not used to fit the model.

iScore,iHessian
Components of the score vector and Hessian matrix for the irregular parameters, if required. See Details.

iArgs
List of extra arguments for the functions iScore, iHessian if required.

Details

Given a fitted spatial point process model model, this function computes the influence measure described in Baddeley, Chang and Song (2013) and Baddeley, Rubak and Turner (2019).

The function influence is generic, and influence.ppm is the method for objects of class "ppm" representing point process models.

The influence of a point process model is a value attached to each data point (i.e. each point of the point pattern to which the model was fitted). The influence value s(x[i]) at a data point x[i] represents the change in the maximised log (pseudo)likelihood that occurs when the point x[i] is deleted. A relatively large value of s(x[i]) indicates a data point with a large influence on the fitted model.

If the point process model trend has irregular parameters that were fitted (using ippm) then the influence calculation requires the first and second derivatives of the log trend with respect to the irregular parameters. The argument iScore should be a list, with one entry for each irregular parameter, of R functions that compute the partial derivatives of the log trend (i.e. log intensity or log conditional intensity) with respect to each irregular parameter. The argument iHessian should be a list, with p^2 entries where p is the number of irregular parameters, of R functions that compute the second order partial derivatives of the log trend with respect to each pair of irregular parameters.

The result of influence.ppm is an object of class "influence.ppm". It can be printed and plotted. It can be converted to a marked point pattern by as.ppp (see as.ppp.influence.ppm). There are also methods for [, as.owin, domain, shift, integral and Smooth.

Value

An object of class "influence.ppm".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

References

Baddeley, A. and Chang, Y.M. and Song, Y. (2013) Leverage and influence diagnostics for spatial point process models. Scandinavian Journal of Statistics 40, 86–104.

Baddeley, A., Rubak, E. and Turner, R. (2019) Leverage and influence diagnostics for Gibbs spatial point processes. Spatial Statistics 29, 15–48.

See Also

leverage.ppm, dfbetas.ppm, ppmInfluence, plot.influence.ppm

Examples

   X <- rpoispp(function(x,y) { exp(3+3*x) })
   fit <- ppm(X ~x+y)
   plot(influence(fit))

[Package spatstat.model version 3.3-4 Index]