integral.msr {spatstat.model}	R Documentation

Integral of a Measure

Description

Computes the integral (total value) of a measure over its domain.

Usage

## S3 method for class 'msr'
integral(f, domain=NULL, weight=NULL, ...)

Arguments

f
A signed measure or vector-valued measure (object of class "msr").

domain
Optional window specifying the domain of integration. Alternatively a tessellation.

weight
Optional. A pixel image (object of class "im") or a function(x,y) giving a numerical weight to be applied to the integration.

...
Ignored.

Details

The integral (total value) of the measure over its domain is calculated.

If domain is a window (class "owin") then the integration will be restricted to this window. If domain is a tessellation (class "tess") then the integral of f in each tile of domain will be computed.

For a multitype measure m, use split.msr to separate the contributions for each type of point, as shown in the Examples.

If weight is given, it should be a pixel image or a function of coordinates x and y returning numerical values. Then each increment of the measure will be multiplied by the corresponding value of weight. Effectively, weight becomes the integrand, and the result is the integral of weight with respect to the measure f.

Value

A numeric value, vector, or matrix.

integral(f) returns a numeric value (for a signed measure) or a vector of values (for a vector-valued measure).

If domain is a tessellation then integral(f, domain) returns a numeric vector with one entry for each tile (if f is a signed measure) or a numeric matrix with one row for each tile (if f is a vector-valued measure).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

msr, integral

Examples

   fit <- ppm(cells ~ x)
   rr <- residuals(fit)
   integral(rr)

   # vector-valued measure
   rs <- residuals(fit, type="score")
   integral(rs)

   # multitype
   fitA <- ppm(amacrine ~ x)
   rrA <- residuals(fitA)
   sapply(split(rrA), integral)

   # multitype and vector-valued
   rsA <- residuals(fitA, type="score")
   sapply(split(rsA), integral)

   ## integral over a subregion
   integral(rr, domain=square(0.5))
   ## integrals over the tiles of a tessellation
   integral(rr, domain=quadrats(cells, 2))

   ## weighted integral
   integral(rr, weight=function(x,y){y})

[Package spatstat.model version 3.3-4 Index]