intensity.dppm {spatstat.model}	R Documentation

Intensity of Determinantal Point Process Model

Description

Extracts the intensity of a determinantal point process model.

Usage

  ## S3 method for class 'detpointprocfamily'
intensity(X, ...)

  ## S3 method for class 'dppm'
intensity(X, ...)

Arguments

X
A determinantal point process model (object of class "detpointprocfamily" or "dppm").

...
Ignored.

Value

A numeric value (if the model is stationary), a pixel image (if the model is non-stationary) or NA if the intensity is unknown for the model.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

[Package spatstat.model version 3.3-4 Index]