intensity.ppm {spatstat.model}	R Documentation

Intensity of Fitted Point Process Model

Description

Computes the intensity of a fitted point process model.

Usage

 ## S3 method for class 'ppm'
intensity(X, ..., approx=c("Poisson", "DPP"))

Arguments

X
A fitted point process model (object of class "ppm").

...
Arguments passed to predict.ppm in some cases. See Details.

approx
Character string (partially matched) specifying the type of approximation to the intensity for a non-Poisson model.

Details

This is a method for the generic function intensity for fitted point process models (class "ppm").

The intensity of a point process model is the expected number of random points per unit area.

If X is a Poisson point process model, the intensity of the process is computed exactly. The result is a numerical value if X is a stationary Poisson point process, and a pixel image if X is non-stationary. (In the latter case, the resolution of the pixel image is controlled by the arguments ... which are passed to predict.ppm.)

If X is a Gibbs point process model that is not a Poisson model, the intensity is computed approximately:

if approx="Poisson" (the default), the intensity is computed using the Poisson-saddlepoint approximation (Baddeley and Nair, 2012a, 2012b, 2017; Anderssen et al, 2014). This approximation is currently available for pairwise-interaction models (Baddeley and Nair, 2012a, 2012b) and for the area-interaction model and Geyer saturation model (Baddeley and Nair, 2017).

If the model is non-stationary. the pseudostationary solution (Baddeley and Nair, 2012b; Anderssen et al, 2014) is used. The result is a pixel image, whose resolution is controlled by the arguments ... which are passed to predict.ppm.

if approx="DPP", the intensity is calculated using the approximation of (Coeurjolly and Lavancier, 2018) based on a determinantal point process. This approximation is more accurate than the Poisson saddlepoint approximation, for inhibitory interactions. However the DPP approximation is only available for stationary pairwise interaction models.

Value

A numeric value (if the model is stationary) or a pixel image.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Gopalan Nair, and Frederic Lavancier.

References

Anderssen, R.S., Baddeley, A., DeHoog, F.R. and Nair, G.M. (2014) Solution of an integral equation arising in spatial point process theory. Journal of Integral Equations and Applications 26 (4) 437–453.

Baddeley, A. and Nair, G. (2012a) Fast approximation of the intensity of Gibbs point processes. Electronic Journal of Statistics 6 1155–1169.

Baddeley, A. and Nair, G. (2012b) Approximating the moments of a spatial point process. Stat 1, 1, 18–30. DOI: 10.1002/sta4.5

Baddeley, A. and Nair, G. (2017) Poisson-saddlepoint approximation for Gibbs point processes with infinite-order interaction: in memory of Peter Hall. Journal of Applied Probability 54, 4, 1008–1026.

Coeurjolly, J.-F. and Lavancier, F. (2018) Approximation intensity for pairwise interaction Gibbs point processes using determinantal point processes. Electronic Journal of Statistics 12 3181–3203.

See Also

intensity, intensity.ppp

Examples

  fitP <- ppm(swedishpines ~ 1)
  intensity(fitP)
  fitS <- ppm(swedishpines ~ 1, Strauss(9))
  intensity(fitS)
  intensity(fitS, approx="D")
  fitSx <- ppm(swedishpines ~ x, Strauss(9))
  lamSx <- intensity(fitSx)
  fitG <- ppm(swedishpines ~ 1, Geyer(9, 1))
  lamG <- intensity(fitG)
  fitA <- ppm(swedishpines ~ 1, AreaInter(7))
  lamA <- intensity(fitA)

[Package spatstat.model version 3.3-4 Index]