intensity.slrm {spatstat.model}	R Documentation

Intensity of Fitted Spatial Logistic Regression Model

Description

Computes the intensity of a fitted spatial logistic regression model, treated as a point process model.

Usage

## S3 method for class 'slrm'
intensity(X, ...)

Arguments

X
A fitted spatial logistic regression model (object of class "slrm").

...
Arguments passed to predict.slrm in some cases. See Details.

Details

This is a method for the generic function intensity for spatial logistic regression models (class "slrm").

The fitted spatial logistic regression model X is interpreted as a point process model. The intensity of a point process model is defined as the expected number of random points per unit area. The fitted probabilities of presence according to X are converted to intensity values.

The result is a numerical value if X is stationary, and a pixel image if X is non-stationary. In the latter case, the resolution of the pixel image is controlled by the arguments ... which are passed to predict.slrm.

Value

A numeric value (if the model is stationary) or a pixel image.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

References

Baddeley, A., Berman, M., Fisher, N.I., Hardegen, A., Milne, R.K., Schuhmacher, D., Shah, R. and Turner, R. (2010) Spatial logistic regression and change-of-support for spatial Poisson point processes. Electronic Journal of Statistics 4, 1151–1201. DOI: 10.1214/10-EJS581

See Also

intensity, intensity.ppm

Examples

  fitS <- slrm(swedishpines ~ 1)
  intensity(fitS)
  fitX <- slrm(swedishpines ~ x)
  intensity(fitX)

[Package spatstat.model version 3.3-4 Index]