interactionorder {spatstat.model}	R Documentation

Determine the Order of Interpoint Interaction in a Model

Description

Given a point process model, report the order of interpoint interaction.

Usage

interactionorder(object)

## S3 method for class 'ppm'
interactionorder(object)

## S3 method for class 'interact'
interactionorder(object)

## S3 method for class 'isf'
interactionorder(object)

## S3 method for class 'fii'
interactionorder(object)

Arguments

object
A point process model (class "ppm") or similar information.

Details

This function determines the order of interpoint interaction in a Gibbs point process model (or a related object).

The interaction order is defined as the largest number k such that the probability density of the model contains terms involving k points at a time. For example, in a pairwise interaction process such as the Strauss process, the probability density contains interaction terms between each pair of points, but does not contain any terms that involve three points at a time, so the interaction order is 2.

Poisson point processes have interaction order 1. Pairwise-interaction processes have interaction order 2. Point processes with the triplet interaction Triplets have interaction order 3. The Geyer saturation model Geyer and the area-interaction model AreaInter have infinite order of interaction.

Value

A positive integer, or Inf.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

Examples

  interactionorder(ppm(cells ~ 1))
  interactionorder(Strauss(0.1))
  interactionorder(Triplets(0.1))
  interactionorder(Geyer(0.1, 2))
  interactionorder(Hybrid(Strauss(0.1), Triplets(0.2)))

[Package spatstat.model version 3.3-4 Index]