is.hybrid {spatstat.model}	R Documentation

Test Whether Object is a Hybrid

Description

Tests where a point process model or point process interaction is a hybrid of several interactions.

Usage

is.hybrid(x)

## S3 method for class 'ppm'
is.hybrid(x)

## S3 method for class 'interact'
is.hybrid(x)

Arguments

x
A point process model (object of class "ppm") or a point process interaction structure (object of class "interact").

Details

A hybrid (Baddeley, Turner, Mateu and Bevan, 2012) is a point process model created by combining two or more point process models, or an interpoint interaction created by combining two or more interpoint interactions.

The function is.hybrid is generic, with methods for point process models (objects of class "ppm") and point process interactions (objects of class "interact"). These functions return TRUE if the object x is a hybrid, and FALSE if it is not a hybrid.

Hybrids of two or more interpoint interactions are created by the function Hybrid. Such a hybrid interaction can then be fitted to point pattern data using ppm.

Value

TRUE if the object is a hybrid, and FALSE otherwise.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

References

Baddeley, A., Turner, R., Mateu, J. and Bevan, A. (2013) Hybrids of Gibbs point process models and their implementation. Journal of Statistical Software 55:11, 1–43. DOI: 10.18637/jss.v055.i11

See Also

Hybrid

Examples

  S <- Strauss(0.1)
  is.hybrid(S)
  H <- Hybrid(Strauss(0.1), Geyer(0.2, 3))
  is.hybrid(H)

  fit <- ppm(redwood, ~1, H)
  is.hybrid(fit)

[Package spatstat.model version 3.3-4 Index]