is.poissonclusterprocess {spatstat.model}	R Documentation

Recognise Poisson Cluster Process Models

Description

Given a point process model (either a model that has been fitted to data, or a model specified by its parameters), determine whether the model is a Poisson cluster process.

Usage

is.poissonclusterprocess(model)
## S3 method for class 'kppm'
is.poissonclusterprocess(model)
## S3 method for class 'zclustermodel'
is.poissonclusterprocess(model)
## Default S3 method:
is.poissonclusterprocess(model)

Arguments

model
Any kind of object representing a spatial point process model, either a model fitted to data, or a specification of a point process model.

Details

The argument model represents a fitted spatial point process model (such as an object of class "ppm", "kppm" or similar) or a specification of a point process model (such as an object of class "zclustermodel").

This function returns TRUE if the model is a Poisson cluster process, and FALSE otherwise.

The function is.poissonclusterprocess is generic, with methods for classes kppm and zclustermodel, and a default method.

Value

A logical value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

kppm, zclustermodel.

Examples

  fut <- kppm(redwood ~ 1, "Thomas")
  is.poissonclusterprocess(fut)
  fot <- slrm(cells ~ x, dimyx=16)
  is.poissonclusterprocess(fot)

[Package spatstat.model version 3.3-4 Index]