is.ppm {spatstat.model}	R Documentation

Test Whether An Object Is A Fitted Point Process Model

Description

Checks whether its argument is a fitted point process model (object of class "ppm", "kppm", "lppm" or "slrm").

Usage

is.ppm(x)
is.kppm(x)
is.lppm(x)
is.slrm(x)

Arguments

x
Any object.

Details

These functions test whether the object x is a fitted point process model object of the specified class.

The result of is.ppm(x) is TRUE if x has "ppm" amongst its classes, and otherwise FALSE. Similarly for the other functions.

Value

A single logical value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

[Package spatstat.model version 3.3-4 Index]