isf.object {spatstat.model}	R Documentation

Interaction Structure Family Objects

Description

Objects of class "isf" are used internally by the spatstat package to represent the structure of the interpoint interactions in a family of point process models.

Details

Advanced Use Only!

An object of class "isf" (Interaction Structure Family) is used internally by the spatstat package to represent the common mathematical and algorithmic structure of the interpoint interactions in a family of point process models.

The existing objects of class "isf" are:

pairwise.family 	 pairwise interaction

triplet.family 	 triplet interaction

pairsat.family 	 saturated pairwise interaction

hierpair.family 	 hierarchical pairwise interaction

inforder.family 	 infinite order interaction

hybrid.family 	 hybrids of several interactions

ord.family 	 Ord interactions

The information contained in these objects enables the spatstat package to select the appropriate algorithm for fitting, predicting and simulating each point process model.

For example, in order to fit a model that involves pairwise interactions, the model-fitting function ppm would use information contained in pairwise.family to select the appropriate algorithms.

An object of class "isf" is essentially a list of functions for various tasks. The internal format is undocumented and may be changed without notice.

Value

An object of class "isf", essentially a list of functions for various tasks.

The internal format is undocumented and may be changed without notice.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

[Package spatstat.model version 3.3-4 Index]