leverage.ppm {spatstat.model}	R Documentation

Leverage Measure for Spatial Point Process Model

Description

Computes the leverage measure for a fitted spatial point process model.

Usage

leverage(model, ...)

## S3 method for class 'ppm'
leverage(model, ...,
         drop = FALSE, iScore=NULL, iHessian=NULL, iArgs=NULL)

Arguments

model
Fitted point process model (object of class "ppm").

...
Ignored, except for the arguments dimyx and eps which are passed to as.mask to control the spatial resolution of the result.

drop
Logical. Whether to include (drop=FALSE) or exclude (drop=TRUE) contributions from quadrature points that were not used to fit the model.

iScore,iHessian
Components of the score vector and Hessian matrix for the irregular parameters, if required. See Details.

iArgs
List of extra arguments for the functions iScore, iHessian if required.

Details

The function leverage is generic, and leverage.ppm is the method for objects of class "ppm".

Given a fitted spatial point process model model, the function leverage.ppm computes the leverage of the model, described in Baddeley, Chang and Song (2013) and Baddeley, Rubak and Turner (2019).

The leverage of a spatial point process model is a function of spatial location, and is typically displayed as a colour pixel image. The leverage value h(u) at a spatial location u represents the change in the fitted trend of the fitted point process model that would have occurred if a data point were to have occurred at the location u. A relatively large value of h() indicates a part of the space where the data have a potentially strong effect on the fitted model (specifically, a strong effect on the intensity or conditional intensity of the fitted model) due to the values of the covariates.

If the point process model trend has irregular parameters that were fitted (using ippm) then the leverage calculation requires the first and second derivatives of the log trend with respect to the irregular parameters. The argument iScore should be a list, with one entry for each irregular parameter, of R functions that compute the partial derivatives of the log trend (i.e. log intensity or log conditional intensity) with respect to each irregular parameter. The argument iHessian should be a list, with p^2 entries where p is the number of irregular parameters, of R functions that compute the second order partial derivatives of the log trend with respect to each pair of irregular parameters.

The result of leverage.ppm is an object of class "leverage.ppm". It can be printed or plotted. It can be converted to a pixel image by as.im (see as.im.leverage.ppm). There are also methods for contour, persp, [, as.function, as.owin, domain, Smooth, integral, and mean.

Value

An object of class "leverage.ppm".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

References

Baddeley, A., Chang, Y.M. and Song, Y. (2013) Leverage and influence diagnostics for spatial point process models. Scandinavian Journal of Statistics 40, 86–104.

Baddeley, A., Rubak, E. and Turner, R. (2019) Leverage and influence diagnostics for Gibbs spatial point processes. Spatial Statistics 29, 15–48.

See Also

influence.ppm, dfbetas.ppm, ppmInfluence, plot.leverage.ppm as.function.leverage.ppm

Examples

if(offline <- !interactive()) op <- spatstat.options(npixel=32, ndummy.min=16)

   X <- rpoispp(function(x,y) { exp(3+3*x) })
   fit <- ppm(X ~x+y)
   le <- leverage(fit)
   if(!offline) plot(le)
   mean(le)

if(offline) spatstat.options(op)

[Package spatstat.model version 3.3-4 Index]