leverage.slrm {spatstat.model}	R Documentation

Leverage and Influence Diagnostics for Spatial Logistic Regression

Description

For a fitted spatial logistic regression model, these functions compute diagnostics of leverage and influence.

Usage

## S3 method for class 'slrm'
leverage(model, ...)
## S3 method for class 'slrm'
influence(model, ...)
## S3 method for class 'slrm'
dfbetas(model, ...)
## S3 method for class 'slrm'
dffit(object, ...)

Arguments

model,object
A fitted spatial logistic regression model (object of class "slrm").

...
Arguments passed to methods.

Details

These functions are methods for the generics leverage, influence, dfbetas and dffit for the class "slrm".

These functions adapt the standard diagnostics for logistic regression (see influence.measures) to a fitted spatial logistic regression model (object of class "slrm"). This adaptation was described by Baddeley, Chang and Song (2013).

leverage.slrm computes the leverage value (diagonal of the hat matrix) for the covariate data in each pixel. The result is a pixel image.

influence.slrm computes the likelihood influence for the data (covariates and presence/absence of points) in each pixel. The result is a pixel image.

dfbetas.slrm computes the parameter influence for the data (covariates and presence/absence of points) in each pixel. The result is a list of pixel images, one image for each of the model coefficients in coef(model). The list can be plotted immediately.

dffit.slrm computes the total influence for the data (covariates and presence/absence of points) in each pixel. The result is a pixel image.

Value

A pixel image, or a list of pixel images.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Baddeley, A., Chang, Y.M. and Song, Y. (2013) Leverage and influence diagnostics for spatial point process models. Scandinavian Journal of Statistics 40, 86–104.

See Also

influence.measures.

leverage.ppm, influence.ppm, dfbetas.ppm, dffit.ppm

Examples

   H <- unmark(humberside)
   fit <- slrm(H ~ x+y, dimyx=32)
   plot(leverage(fit))
   plot(influence(fit))
   plot(dfbetas(fit))
   plot(dffit(fit))

[Package spatstat.model version 3.3-4 Index]