logLik.dppm {spatstat.model}	R Documentation

Log Likelihood and AIC for Fitted Determinantal Point Process Model

Description

Extracts the log Palm likelihood, deviance, and AIC of a fitted determinantal point process model.

Usage

## S3 method for class 'dppm'
logLik(object, ...)
## S3 method for class 'dppm'
AIC(object, ..., k=2)
## S3 method for class 'dppm'
extractAIC(fit, scale=0, k=2, ...)
## S3 method for class 'dppm'
nobs(object, ...)

Arguments

object,fit
Fitted point process model. An object of class "dppm".

...
Ignored.

scale
Ignored.

k
Numeric value specifying the weight of the equivalent degrees of freedom in the AIC. See Details.

Details

These functions are methods for the generic commands logLik, extractAIC and nobs for the class "dppm".

An object of class "dppm" represents a fitted Cox or cluster point process model. It is obtained from the model-fitting function dppm.

These methods apply only when the model was fitted by maximising the Palm likelihood (Tanaka et al, 2008) by calling dppm with the argument method="palm".

The method logLik.dppm computes the maximised value of the log Palm likelihood for the fitted model object.

The methods AIC.dppm and extractAIC.dppm compute the Akaike Information Criterion AIC for the fitted model based on the Palm likelihood (Tanaka et al, 2008)

AIC = -2 * log(PL) + k * edf

where PL is the maximised Palm likelihood of the fitted model, and edf is the effective degrees of freedom of the model.

The method nobs.dppm returns the number of points in the original data point pattern to which the model was fitted.

The R function step uses these methods, but it does not work for determinantal models yet due to a missing implementation of update.dppm.

Value

logLik returns a numerical value, belonging to the class "logLik", with an attribute "df" giving the degrees of freedom.

AIC returns a numerical value.

extractAIC returns a numeric vector of length 2 containing the degrees of freedom and the AIC value.

nobs returns an integer value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

References

Tanaka, U. and Ogata, Y. and Stoyan, D. (2008) Parameter estimation and model selection for Neyman-Scott point processes. Biometrical Journal 50, 43–57.

See Also

dppm, logLik.ppm

Examples

  fit <- dppm(swedishpines ~ x, dppGauss(), method="palm")
  nobs(fit)
  logLik(fit)
  extractAIC(fit)
  AIC(fit)

[Package spatstat.model version 3.3-4 Index]