logLik.kppm {spatstat.model}	R Documentation

Log Likelihood and AIC for Fitted Cox or Cluster Point Process Model

Description

Extracts the log composite likelihood, deviance, and AIC of a fitted Cox or cluster point process model.

Usage

## S3 method for class 'kppm'
logLik(object, ...)
## S3 method for class 'kppm'
AIC(object, ..., k=2)
## S3 method for class 'kppm'
extractAIC(fit, scale=0, k=2, ...)
## S3 method for class 'kppm'
nobs(object, ...)

Arguments

object,fit
Fitted point process model. An object of class "kppm".

...
Ignored.

scale
Ignored.

k
Numeric value specifying the weight of the equivalent degrees of freedom in the AIC. See Details.

Details

These functions are methods for the generic commands logLik, extractAIC and nobs for the class "kppm".

An object of class "kppm" represents a fitted Cox or cluster point process model. It is obtained from the model-fitting function kppm.

These methods apply only when the model was fitted by maximising a composite likelihood: either the Palm likelihood (Tanaka et al, 2008) or the second order composite likelihood (Guan, 2006), by calling kppm with the argument method="palm" or method="clik2" respectively.

The method logLik.kppm computes the maximised value of the log composite likelihood for the fitted model object.

The methods AIC.kppm and extractAIC.kppm compute the Akaike Information Criterion AIC for the fitted model based on the composite likelihood

AIC = -2 * log(CL) + k * edf

where CL is the maximised composite likelihood of the fitted model, and edf is the effective degrees of freedom of the model.

The method nobs.kppm returns the number of points in the original data point pattern to which the model was fitted.

The R function step uses these methods.

Value

logLik returns a numerical value, belonging to the class "logLik", with an attribute "df" giving the degrees of freedom.

AIC returns a numerical value.

extractAIC returns a numeric vector of length 2 containing the degrees of freedom and the AIC value.

nobs returns an integer value.

Model comparison

The values of log-likelihood and AIC returned by these functions are based on the composite likelihood of the cluster process or Cox process model. They are available only when the model was fitted using method="palm" or method="clik2".

For model comparison and model selection, it is valid to compare the logLik values, or to compare the AIC values, but only when all the models are of class "kppm" and were fitted using the same method.

For method="palm" some theoretical justification was provided by Tanaka et al (2008).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

References

Guan, Y. (2006) A composite likelihood approach in fitting spatial point process models. Journal of the American Statistical Association 101, 1502–1512.

Tanaka, U. and Ogata, Y. and Stoyan, D. (2008) Parameter estimation and model selection for Neyman-Scott point processes. Biometrical Journal 50, 43–57.

See Also

kppm, logLik.ppm

Examples

  fit <- kppm(redwood ~ x, "Thomas", method="palm")
  nobs(fit)
  logLik(fit)
  extractAIC(fit)
  AIC(fit)
  step(fit)

[Package spatstat.model version 3.3-4 Index]