measureContinuous {spatstat.model}	R Documentation

Discrete and Continuous Components of a Measure

Description

Given a measure A (object of class "msr") these functions find the discrete and continuous parts of A.

Usage

measureDiscrete(x)
measureContinuous(x)

Arguments

x
A measure (object of class "msr").

Details

The functions measureDiscrete and measureContinuous return the discrete and continuous components, respectively, of a measure.

If x is a measure, then measureDiscrete(x) is a measure consisting only of the discrete (atomic) component of x, and measureContinuous(x) is a measure consisting only of the continuous (diffuse) component of x.

Value

Another measure (object of class "msr") on the same spatial domain.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Halmos, P.R. (1950) Measure Theory. Van Nostrand.

See Also

msr, with.msr, split.msr, measurePositive

Examples

   X <- rpoispp(function(x,y) { exp(3+3*x) })
   fit <- ppm(X, ~x+y)
   rp <- residuals(fit, type="pearson")

   rp
   measureDiscrete(rp)
   measureContinuous(rp)

[Package spatstat.model version 3.3-4 Index]