measureVariation {spatstat.model}	R Documentation

Positive and Negative Parts, and Variation, of a Measure

Description

Given a measure A (object of class "msr") these functions find the positive part, negative part and variation of A.

Usage

measurePositive(x)
measureNegative(x)
measureVariation(x)
totalVariation(x)

Arguments

x
A measure (object of class "msr").

Details

The functions measurePositive and measureNegative return the positive and negative parts of the measure, and measureVariation returns the variation (sum of positive and negative parts). The function totalVariation returns the total variation norm.

If μ is a signed measure, it can be represented as

μ = μ[+] - μ[-]

where μ[+] and μ[-] are nonnegative measures called the positive and negative parts of μ. In a nutshell, the positive part of μ consists of all positive contributions or increments, and the negative part consists of all negative contributions multiplied by -1.

The variation |μ| is defined by

μ = μ[+] + μ[-]

and is also a nonnegative measure.

The total variation norm is the integral of the variation.

Value

The result of measurePositive, measureNegative and measureVariation is another measure (object of class "msr") on the same spatial domain. The result of totalVariation is a non-negative number.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Halmos, P.R. (1950) Measure Theory. Van Nostrand.

See Also

msr, with.msr, split.msr, measureDiscrete

Examples

   X <- rpoispp(function(x,y) { exp(3+3*x) })
   fit <- ppm(X, ~x+y)
   rp <- residuals(fit, type="pearson")

   measurePositive(rp)
   measureNegative(rp)
   measureVariation(rp)

   # total variation norm
   totalVariation(rp)

[Package spatstat.model version 3.3-4 Index]