measureWeighted {spatstat.model}	R Documentation

Weighted Version of a Measure

Description

Given a measure m (object of class "msr") and a spatially-varying weight function, construct the weighted version of m.

Usage

measureWeighted(m, w)

Arguments

m
A measure (object of class "msr").

w
A pixel image (object of class "im") or a function(x,y) giving the numeric weight at each spatial location.

Details

For any region of space B, the weighted measure wm has the value

wm(B) = integral[B] w(x) dm(x)

In any small region of space, the increment of the weighted measure wm is equal to the increment of the original measure m multiplied by the weight w at that location.

Value

Another measure (object of class "msr") on the same spatial domain.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Halmos, P.R. (1950) Measure Theory. Van Nostrand.

See Also

msr, with.msr, split.msr, measurePositive

Examples

   fit <- ppm(cells ~ x)
   res <- residuals(fit)
   measureWeighted(res, function(x,y){x})

[Package spatstat.model version 3.3-4 Index]