methods.dppm {spatstat.model}	R Documentation

Methods for Determinantal Point Process Models

Description

These are methods for the class "dppm".

Usage

## S3 method for class 'dppm'
coef(object, ...)
## S3 method for class 'dppm'
formula(x, ...)
## S3 method for class 'dppm'
print(x, ...)
## S3 method for class 'dppm'
terms(x, ...)
## S3 method for class 'dppm'
labels(object, ...)

Arguments

x,object
An object of class "dppm", representing a fitted determinantal point process model.

...
Arguments passed to other methods.

Details

These functions are methods for the generic commands coef, formula, print, terms and labels for the class "dppm".

An object of class "dppm" represents a fitted determinantal point process model. It is obtained from dppm.

The method coef.dppm returns the vector of regression coefficients of the fitted model. It does not return the interaction parameters.

Value

See the help files for the corresponding generic functions.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

dppm, plot.dppm, predict.dppm, simulate.dppm, as.ppm.dppm.

Examples

  fit <- dppm(swedishpines ~ x + y, dppGauss, method="c")
  coef(fit)
  formula(fit)
  tf <- terms(fit)
  labels(fit)

[Package spatstat.model version 3.3-4 Index]