methods.fii {spatstat.model}	R Documentation

Methods for Fitted Interactions

Description

These are methods specifically for the class "fii" of fitted interpoint interactions.

Usage

## S3 method for class 'fii'
print(x, ...)

## S3 method for class 'fii'
coef(object, ...)

## S3 replacement method for class 'fii'
coef(object, ...) <- value

## S3 method for class 'fii'
plot(x, ...)

## S3 method for class 'fii'
summary(object,...)

## S3 method for class 'summary.fii'
print(x, ...)

## S3 method for class 'summary.fii'
coef(object, ...)

Arguments

x,object
An object of class "fii" representing a fitted interpoint interaction.

...
Arguments passed to other methods.

value
Numeric vector containing new values for the fitted interaction coefficients.

Details

These are methods for the class "fii". An object of class "fii" represents a fitted interpoint interaction. It is usually obtained by using the command fitin to extract the fitted interaction part of a fitted point process model. See fitin for further explanation of this class.

The commands listed here are methods for the generic functions print, summary, plot, coef and coef<- for objects of the class "fii".

Following the usual convention, summary.fii returns an object of class summary.fii, for which there is a print method. The effect is that, when the user types summary(x), the summary is printed, but when the user types y <- summary(x), the summary information is saved.

The method coef.fii extracts the canonical coefficients of the fitted interaction, and returns them as a numeric vector. The method coef.summary.fii transforms these values into quantities that are more easily interpretable, in a format that depends on the particular model.

There are also methods for the generic commands reach and as.interact, described elsewhere.

Value

The print and plot methods return NULL.

The summary method returns an object of class summary.fii.

coef.fii returns a numeric vector. coef.summary.fii returns data whose structure depends on the model.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

fitin, reach.fii, as.interact.fii

Examples

  mod <- ppm(cells ~1, Strauss(0.1))
  f <- fitin(mod)
  f
  summary(f)
  plot(f)
  coef(f)
  coef(summary(f))

[Package spatstat.model version 3.3-4 Index]