methods.influence.ppm {spatstat.model}	R Documentation

Methods for Influence Objects

Description

Methods for the class "influence.ppm".

Usage

  ## S3 method for class 'influence.ppm'
as.ppp(X, ...)

 ## S3 method for class 'influence.ppm'
as.owin(W, ..., fatal=TRUE)

 ## S3 method for class 'influence.ppm'
domain(X, ...)

 ## S3 method for class 'influence.ppm'
Smooth(X, ...)

 ## S3 method for class 'influence.ppm'
Window(X, ...)

 ## S3 method for class 'influence.ppm'
integral(f, domain, ...)

Arguments

X,W,f
An object of class "influence.ppm".

domain
Optional. Domain of integration: a window (class "owin") or a tessellation (class "tess").

...
Additional arguments. See Details.

fatal
Logical value indicating what to do if the data cannot be converted to a window. If fatal=TRUE (the default) an error occurs. If fatal=FALSE a value of NULL is returned.

Details

These functions are methods for the class "influence.ppm". An object of this class represents the influence measure of a fitted point process model (see influence.ppm).

For as.ppp, domain, integral and Window, additional arguments (...) are ignored. For as.owin and Smooth, additional arguments are passed to the method for class "ppp".

Value

A window (object of class "owin") for as.owin, domain and Window. A point pattern (object of class "ppp") for as.ppp. A numeric value or numeric vector for integral. A pixel image, or list of pixel images, for Smooth.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

influence.ppm, plot.influence.ppm, [.influence.ppm

Examples

 fit <- ppm(cells ~ x)
 a <- influence(fit)
 Window(a)

[Package spatstat.model version 3.3-4 Index]