methods.kppm {spatstat.model}	R Documentation

Methods for Cluster Point Process Models

Description

These are methods for the class "kppm".

Usage

## S3 method for class 'kppm'
coef(object, ...)
## S3 method for class 'kppm'
formula(x, ...)
## S3 method for class 'kppm'
print(x, ...)
## S3 method for class 'kppm'
terms(x, ...)
## S3 method for class 'kppm'
labels(object, ...)

Arguments

x,object
An object of class "kppm", representing a fitted cluster point process model.

...
Arguments passed to other methods.

Details

These functions are methods for the generic commands coef, formula, print, terms and labels for the class "kppm".

An object of class "kppm" represents a fitted cluster point process model. It is obtained from kppm.

The method coef.kppm returns the vector of regression coefficients of the fitted model. It does not return the clustering parameters.

Value

See the help files for the corresponding generic functions.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

kppm, plot.kppm, predict.kppm, simulate.kppm, update.kppm, vcov.kppm, as.ppm.kppm.

Examples

  fit <- kppm(redwood ~ x, "MatClust")
  coef(fit)
  formula(fit)
  tf <- terms(fit)
  labels(fit)

[Package spatstat.model version 3.3-4 Index]