methods.leverage.ppm {spatstat.model}	R Documentation

Methods for Leverage Objects

Description

Methods for the class "leverage.ppm".

Usage

  ## S3 method for class 'leverage.ppm'
as.im(X, ..., what=c("smooth", "nearest"))

 ## S3 method for class 'leverage.ppm'
as.owin(W, ..., fatal=TRUE)

 ## S3 method for class 'leverage.ppm'
domain(X, ...)

 ## S3 method for class 'leverage.ppm'
integral(f, domain, ...)

 ## S3 method for class 'leverage.ppm'
mean(x, ...)

 ## S3 method for class 'leverage.ppm'
Smooth(X, ...)

 ## S3 method for class 'leverage.ppm'
Window(X, ...)

Arguments

X,x,W,f
An object of class "leverage.ppm".

domain
Optional. Domain of integration: a window (class "owin") or a tessellation (class "tess").

...
Additional arguments. See Details.

fatal
Logical value indicating what to do if the data cannot be converted to a window. If fatal=TRUE (the default) an error occurs. If fatal=FALSE a value of NULL is returned.

what
Character string (partially matched) specifying which image data should be extracted. See plot.leverage.ppm for explanation.

Details

These functions are methods for the class "leverage.ppm". An object of this class represents the leverage measure of a fitted point process model (see leverage.ppm).

For as.im, domain and Window, additional arguments (...) are ignored. For as.owin, integral, mean and Smooth, additional arguments are passed to the method for class "im".

Value

A window (object of class "owin") for as.owin, domain and Window. A numeric value or numeric vector for integral. A pixel image, or list of pixel images, for as.im and Smooth.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

leverage.ppm, plot.leverage.ppm, [.leverage.ppm, as.function.leverage.ppm.

Examples

 fit <- ppm(cells ~ x)
 a <- leverage(fit)
 integral(a)

[Package spatstat.model version 3.3-4 Index]