methods.objsurf {spatstat.model}	R Documentation

Methods for Objective Function Surfaces

Description

Methods for printing and plotting an objective function surface.

Usage

## S3 method for class 'objsurf'
print(x, ...)
## S3 method for class 'objsurf'
plot(x, ...)
## S3 method for class 'objsurf'
image(x, ...)
## S3 method for class 'objsurf'
contour(x, ...)
## S3 method for class 'objsurf'
persp(x, ...)
## S3 method for class 'objsurf'
summary(object, ...)
## S3 method for class 'summary.objsurf'
print(x, ...)

Arguments

x,object
Object of class "objsurf" representing an objective function surface.

...
Additional arguments passed to plot methods.

Details

These are methods for the generic functions print, plot, image, contour, persp and summary for the class "objsurf".

Value

For print.objsurf, print.summary.objsurf, plot.objsurf and image.objsurf the value is NULL.

For contour.objsurf and persp.objsurf the value is described in the help for contour.default and persp.default respectively.

For summary.objsurf the result is a list, of class summary.objsurf, containing summary information. This list is printed in sensible format by print.summary.objsurf.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Ege Rubak rubak@math.aau.dk.

See Also

objsurf

Examples

 fit <- kppm(redwood ~ 1, "Thomas")
 os <- objsurf(fit)
  os
  summary(os)
  plot(os)
  contour(os, add=TRUE)
  persp(os)

[Package spatstat.model version 3.3-4 Index]