methods.slrm {spatstat.model}	R Documentation

Methods for Spatial Logistic Regression Models

Description

These are methods for the class "slrm".

Usage

## S3 method for class 'slrm'
formula(x, ...)
## S3 method for class 'slrm'
print(x, ...)
## S3 method for class 'slrm'
summary(object, ...)
## S3 method for class 'slrm'
terms(x, ...)
## S3 method for class 'slrm'
labels(object, ...)
## S3 method for class 'slrm'
deviance(object, ...)
## S3 method for class 'slrm'
update(object, fmla, ..., evaluate = TRUE, env = parent.frame())

Arguments

x,object
An object of class "slrm", representing a fitted spatial logistic regression model.

...
Arguments passed to other methods.

fmla
Optional. A formula, to replace the formula of the model.

evaluate
Logical value. If TRUE, evaluate the updated call to slrm, so that the model is refitted; if FALSE, simply return the updated call.

env
Optional environment in which the model should be updated.

Details

These functions are methods for the generic commands formula, update, print, summary, terms, labels and deviance for the class "slrm".

An object of class "slrm" represents a fitted spatial logistic regression model. It is obtained from slrm.

Value

See the help files for the corresponding generic functions.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

slrm, plot.slrm, predict.slrm, simulate.slrm, vcov.slrm, coef.slrm.

Examples

  fit <- slrm(redwood ~ x)
  coef(fit)
  formula(fit)
  tf <- terms(fit)  
  labels(fit)
  deviance(fit) 

[Package spatstat.model version 3.3-4 Index]