methods.traj {spatstat.model}	R Documentation

Methods for Trajectories of Function Evaluations

Description

Methods for objects of class "traj".

Usage

## S3 method for class 'traj'
print(x, ...)
## S3 method for class 'traj'
plot(x, ..., show.ends=TRUE, add=FALSE, xlab=NULL, ylab=NULL)
## S3 method for class 'traj'
lines(x, ..., directed=FALSE)

Arguments

x
Object of class "traj".

...
Additional arguments passed to other methods.

directed
Logical value specifying whether to draw arrows instead of undirected lines.

show.ends
Logical value specifying whether to indicate the start and finish of the trajectory. The start is a blue circle; the finish is a red cross.

add
Logical value specifying whether to draw the trajectory on the existing plot (add=TRUE) or to start a new plot (add=FALSE, the default).

xlab,ylab
Optional labels for the horizontal and vertical axes.

Details

An object of class "traj" represents the history of evaluations of the objective function performed when a cluster process model was fitted. It is a data frame containing the input parameter values for the objective function, and the corresponding value of the objective function, that were considered by the optimisation algorithm.

These functions are methods for the generic print, plot and lines.

Value

Null.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

traj

Examples

  fit <- kppm(redwood, pspace=list(save=TRUE))
  h <- traj(fit)
  h
  plot(h)
  lines(h)

[Package spatstat.model version 3.3-4 Index]