methods.zclustermodel {spatstat.model}	R Documentation

Methods for Cluster Models

Description

Methods for the experimental class of cluster models.

Usage

 ## S3 method for class 'zclustermodel'
pcfmodel(model, ...)

 ## S3 method for class 'zclustermodel'
Kmodel(model, ...)

 ## S3 method for class 'zclustermodel'
intensity(X, ...)

 ## S3 method for class 'zclustermodel'
predict(object, ...,
                  locations, type = "intensity", ngrid = NULL)

 ## S3 method for class 'zclustermodel'
print(x, ...)

 ## S3 method for class 'zclustermodel'
clusterradius(model,...,thresh=NULL, precision=FALSE)

 ## S3 method for class 'zclustermodel'
reach(x, ..., epsilon)

Arguments

model,object,x,X
Object of class "zclustermodel".

...
Arguments passed to other methods.

locations
Locations where prediction should be performed. A window or a point pattern.

type
Currently must equal "intensity".

ngrid
Pixel grid dimensions for prediction, if locations is a rectangle or polygon.

thresh,epsilon
Tolerance thresholds

precision
Logical value stipulating whether the precision should also be returned.

Details

Experimental.

Value

Same as for other methods.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

zclustermodel

Examples

  m <- zclustermodel("Thomas", kappa=10, mu=5, scale=0.1)
  m2 <- zclustermodel("VarGamma", kappa=10, mu=10, scale=0.1, nu=0.7)
  m
  m2
  g <- pcfmodel(m)
  g(0.2)
  g2 <- pcfmodel(m2)
  g2(1)
  Z <- predict(m, locations=square(2))
  Z2 <- predict(m2, locations=square(1))
  varcount(m, square(1))
  varcount(m2, square(1))

[Package spatstat.model version 3.3-4 Index]