mppm {spatstat.model}	R Documentation

Fit Point Process Model to Several Point Patterns

Description

Fits a Gibbs point process model to several point patterns simultaneously.

Usage

   mppm(formula, data, interaction=Poisson(), ...,
        iformula=NULL, 
        random=NULL,
        weights=NULL,
        use.gam = FALSE, 
        reltol.pql=1e-3,
        gcontrol=list())

Arguments

formula
A formula describing the systematic part of the model. Variables in the formula are names of columns in data.

data
A hyperframe (object of class "hyperframe", see hyperframe) containing the point pattern responses and the explanatory variables.

interaction
Interpoint interaction(s) appearing in the model. Either an object of class "interact" describing the point process interaction structure, or a hyperframe (with the same number of rows as data) whose entries are objects of class "interact".

...
Arguments passed to ppm controlling the fitting procedure.

iformula
Optional. A formula (with no left hand side) describing the interaction to be applied to each case. Each variable name in the formula should either be the name of a column in the hyperframe interaction, or the name of a column in the hyperframe data that is a vector or factor.

random
Optional. A formula (with no left hand side) describing a random effect. Variable names in the formula may be any of the column names of data and interaction. The formula must be recognisable to lme.

weights
Optional. Numeric vector of case weights for each row of data.

use.gam
Logical flag indicating whether to fit the model using gam or glm.

reltol.pql
Relative tolerance for successive steps in the penalised quasi-likelihood algorithm, used when the model includes random effects. The algorithm terminates when the root mean square of the relative change in coefficients is less than reltol.pql.

gcontrol
List of arguments to control the fitting algorithm. Arguments are passed to glm.control or gam.control or lmeControl depending on the kind of model being fitted. If the model has random effects, the arguments are passed to lmeControl. Otherwise, if use.gam=TRUE the arguments are passed to gam.control, and if use.gam=FALSE (the default) they are passed to glm.control.

Details

This function fits a common point process model to a dataset containing several different point patterns.

It extends the capabilities of the function ppm to deal with data such as

replicated observations of spatial point patterns

two groups of spatial point patterns

a designed experiment in which the response from each unit is a point pattern.

The syntax of this function is similar to that of standard R model-fitting functions like lm and glm. The first argument formula is an R formula describing the systematic part of the model. The second argument data contains the responses and the explanatory variables. Other arguments determine the stochastic structure of the model.

Schematically, the data are regarded as the results of a designed experiment involving n experimental units. Each unit has a ‘response’, and optionally some ‘explanatory variables’ (covariates) describing the experimental conditions for that unit. In this context, the response from each unit is a point pattern. The value of a particular covariate for each unit can be either a single value (numerical, logical or factor), or a spatial covariate. A ‘spatial’ covariate is a quantity that depends on spatial location, for example, the soil acidity or altitude at each location. For the purposes of mppm, a spatial covariate must be stored as a pixel image (object of class "im") which gives the values of the covariate at a fine grid of locations.

The argument data is a hyperframe (a generalisation of a data frame, see hyperframe). This is like a data frame except that the entries can be objects of any class. The hyperframe has one row for each experimental unit, and one column for each variable (response or explanatory variable).

The formula should be an R formula. The left hand side of formula determines the ‘response’ variable. This should be a single name, which should correspond to a column in data.

The right hand side of formula determines the spatial trend of the model. It specifies the linear predictor, and effectively represents the logarithm of the spatial trend. Variables in the formula must be the names of columns of data, or one of the reserved names

x,y

Cartesian coordinates of location

marks

Mark attached to point

id

which is a factor representing the serial number (1 to n) of the point pattern, i.e. the row number in the data hyperframe.

The column of responses in data must consist of point patterns (objects of class "ppp"). The individual point pattern responses can be defined in different spatial windows. If some of the point patterns are marked, then they must all be marked, and must have the same type of marks.

The scope of models that can be fitted to each pattern is the same as the scope of ppm, that is, Gibbs point processes with interaction terms that belong to a specified list, including for example the Poisson process, Strauss process, Geyer's saturation model, and piecewise constant pairwise interaction models. Additionally, it is possible to include random effects as explained in the section on Random Effects below.

The stochastic part of the model is determined by the arguments interaction and (optionally) iformula.

In the simplest case, interaction is an object of class "interact", determining the interpoint interaction structure of the point process model, for all experimental units.

Alternatively, interaction may be a hyperframe, whose entries are objects of class "interact". It should have the same number of rows as data.

If interaction consists of only one column, then the entry in row i is taken to be the interpoint interaction for the ith experimental unit (corresponding to the ith row of data).

If interaction has more than one column, then the argument iformula is also required. Each row of interaction determines several interpoint interaction structures that might be applied to the corresponding row of data. The choice of interaction is determined by iformula; this should be an R formula, without a left hand side. For example if interaction has two columns called A and B then iformula = ~B indicates that the interpoint interactions are taken from the second column.

Variables in iformula typically refer to column names of interaction. They can also be names of columns in data, but only for columns of numeric, logical or factor values. For example iformula = ~B * group (where group is a column of data that contains a factor) causes the model with interpoint interaction B to be fitted with different interaction parameters for each level of group.

Value

An object of class "mppm" representing the fitted model.

There are methods for print, summary, coef, AIC, anova, fitted, fixef, logLik, plot, predict, ranef, residuals, summary, terms and vcov for this class.

The default methods for update and formula also work on this class.

Random Effects

It is also possible to include random effects in the trend term. The argument random is a formula, with no left-hand side, that specifies the structure of the random effects. The formula should be recognisable to lme (see the description of the argument random for lme).

The names in the formula random may be any of the covariates supplied by data. Additionally the formula may involve the name id, which is a factor representing the serial number (1 to n) of the point pattern in the list X.

Author(s)

Adrian Baddeley, Ida-Maria Sintorn and Leanne Bischoff. Implemented in spatstat by Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

References

Baddeley, A. and Turner, R. Practical maximum pseudolikelihood for spatial point patterns. Australian and New Zealand Journal of Statistics 42 (2000) 283–322.

Baddeley, A., Bischof, L., Sintorn, I.-M., Haggarty, S., Bell, M. and Turner, R. Analysis of a designed experiment where the response is a spatial point pattern. In preparation.

Baddeley, A., Rubak, E. and Turner, R. (2015) Spatial Point Patterns: Methodology and Applications with R. Chapman and Hall/CRC Press.

Bell, M. and Grunwald, G. (2004) Mixed models for the analysis of replicated spatial point patterns. Biostatistics 5, 633–648.

See Also

ppm, print.mppm, summary.mppm, coef.mppm,

Examples

# Waterstriders data
 H <- hyperframe(Y = waterstriders)
 mppm(Y ~ 1,  data=H)
 mppm(Y ~ 1,  data=H, Strauss(7))
 mppm(Y ~ id, data=H)
 mppm(Y ~ x,  data=H)

# Synthetic data from known model
n <- 10
H <- hyperframe(V=1:n,
                U=runif(n, min=-1, max=1),
                M=factor(letters[1 + (1:n) %% 3]))
H$Z <- setcov(square(1))
H$U <- with(H, as.im(U, as.rectangle(Z)))
H$Y <- with(H, rpoispp(eval.im(exp(2+3*Z))))

fit <- mppm(Y ~Z + U + V, data=H)

[Package spatstat.model version 3.3-4 Index]