npfun {spatstat.model}	R Documentation

Dummy Function Returns Number of Points

Description

Returns a summary function which is constant with value equal to the number of points in the point pattern.

Usage

npfun(X, ..., r)

Arguments

X
Point pattern.

...
Ignored.

r
Vector of values of the distance argument r.

Details

This function is normally not called by the user. Instead it is passed as an argument to the function psst.

Value

Object of class "fv" representing a constant function.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Ege Rubak rubak@math.aau.dk and Jesper Moller.

See Also

psst

Examples

  fit0 <- ppm(cells, ~1, nd=10)
  v <- psst(fit0, npfun)

[Package spatstat.model version 3.3-4 Index]