pairsat.family {spatstat.model}	R Documentation

Saturated Pairwise Interaction Point Process Family

Description

An object describing the Saturated Pairwise Interaction family of point process models

Details

Advanced Use Only!

This structure would not normally be touched by the user. It describes the “saturated pairwise interaction” family of point process models.

If you need to create a specific interaction model for use in spatial pattern analysis, use the function Saturated() or the two existing implementations of models in this family, Geyer() and SatPiece().

Geyer (1999) introduced the “saturation process”, a modification of the Strauss process in which the total contribution to the potential from each point (from its pairwise interaction with all other points) is trimmed to a maximum value c. This model is implemented in the function Geyer().

The present class pairsat.family is the extension of this saturation idea to all pairwise interactions. Note that the resulting models are no longer pairwise interaction processes - they have interactions of infinite order.

pairsat.family is an object of class "isf" containing a function pairwise$eval for evaluating the sufficient statistics of any saturated pairwise interaction point process model in which the original pair potentials take an exponential family form.

Value

Object of class "isf", see isf.object.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

References

Geyer, C.J. (1999) Likelihood Inference for Spatial Point Processes. Chapter 3 in O.E. Barndorff-Nielsen, W.S. Kendall and M.N.M. Van Lieshout (eds) Stochastic Geometry: Likelihood and Computation, Chapman and Hall / CRC, Monographs on Statistics and Applied Probability, number 80. Pages 79–140.

See Also

Geyer to create the Geyer saturation process.

SatPiece to create a saturated process with piecewise constant pair potential.

Saturated to create a more general saturation model.

Other families: inforder.family, ord.family, pairwise.family.

[Package spatstat.model version 3.3-4 Index]