pairwise.family {spatstat.model}	R Documentation

Pairwise Interaction Process Family

Description

An object describing the family of all pairwise interaction Gibbs point processes.

Details

Advanced Use Only!

This structure would not normally be touched by the user. It describes the pairwise interaction family of point process models.

If you need to create a specific pairwise interaction model for use in modelling, use the function Pairwise or one of the existing functions listed below.

Anyway, pairwise.family is an object of class "isf" containing a function pairwise.family$eval for evaluating the sufficient statistics of any pairwise interaction point process model taking an exponential family form.

Value

Object of class "isf", see isf.object.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

Other families: pairsat.family, ord.family, inforder.family.

Pairwise interactions: Poisson, Pairwise, PairPiece, Fiksel, Hardcore, LennardJones, MultiHard, MultiStrauss, MultiStraussHard, Strauss, StraussHard, Softcore.

Other interactions: AreaInter, Geyer, Saturated, Ord, OrdThresh.

[Package spatstat.model version 3.3-4 Index]