panysib {spatstat.model}	R Documentation

Probability that a Point Has Any Siblings

Description

Given a cluster process model, calculate the probability that a point of the process has any siblings.

Usage

panysib(object)

Arguments

object
Fitted cluster process model (object of class "kppm").

Details

In a Poisson cluster process, two points are called siblings if they belong to the same cluster, that is, if they had the same parent point. This function computes the probability that a given random point has any siblings.

If object is a stationary point process, the result is a single number, which is the probability that a typical point of the process has any siblings. If this number is small, then the process is approximately a homogeneous Poisson process (complete spatial randomness). The converse is not true (Baddeley et al, 2022).

Otherwise, the result is a pixel image, in which the value at any location u is the conditional probability, given there is a point of the process at u, that this point has any siblings. If the pixel values are all small, then the process is approximately an inhomogeneous Poisson process.

This concept was proposed by Baddeley et al (2022).

Value

A single number (if object is a stationary point process) or a pixel image (otherwise).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Baddeley, A., Davies, T.M., Hazelton, M.L., Rakshit, S. and Turner, R. (2022) Fundamental problems in fitting spatial cluster process models. Spatial Statistics 52, 100709. DOI: 10.1016/j.spasta.2022.100709

See Also

psib

Examples

  fit <- kppm(redwood ~ polynom(x,y,2))
  plot(panysib(fit))

[Package spatstat.model version 3.3-4 Index]