parameters {spatstat.model}	R Documentation

Extract Model Parameters in Understandable Form

Description

Given a fitted model of some kind, this function extracts all the parameters needed to specify the model, and returns them as a list.

Usage

parameters(model, ...)

## S3 method for class 'dppm'
parameters(model, ...)

## S3 method for class 'kppm'
parameters(model, ...)

## S3 method for class 'slrm'
parameters(model, ...)

## S3 method for class 'ppm'
parameters(model, ...)

## S3 method for class 'profilepl'
parameters(model, ...)

## S3 method for class 'fii'
parameters(model, ...)

## S3 method for class 'interact'
parameters(model, ...)

Arguments

model
A fitted model of some kind.

...
Arguments passed to methods.

Details

The argument model should be a fitted model of some kind. This function extracts all the parameters that would be needed to specify the model, and returns them as a list.

The function parameters is generic, with methods for class "ppm", "kppm", "dppm" and "profilepl" and other classes.

Value

A named list, whose format depends on the fitted model.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

coef

Examples

  parameters(Strauss(0.1))
  fit1 <- ppm(cells ~ x, Strauss(0.1))
  parameters(fit1)
  fit2 <- kppm(redwood ~ x, "Thomas")
  parameters(fit2)

[Package spatstat.model version 3.3-4 Index]