plot.dppm {spatstat.model}	R Documentation

Plot a fitted determinantal point process

Description

Plots a fitted determinantal point process model, displaying the fitted intensity and the fitted summary function.

Usage

  ## S3 method for class 'dppm'
plot(x, ..., what=c("intensity", "statistic"))

Arguments

x
Fitted determinantal point process model. An object of class "dppm".

...
Arguments passed to plot.ppm and plot.fv to control the plot.

what
Character vector determining what will be plotted.

Details

This is a method for the generic function plot for the class "dppm" of fitted determinantal point process models.

The argument x should be a determinantal point process model (object of class "dppm") obtained using the function dppm.

The choice of plots (and the order in which they are displayed) is controlled by the argument what. The options (partially matched) are "intensity" and "statistic".

This command is capable of producing two different plots:

what="intensity"

specifies the fitted intensity of the model, which is plotted using plot.ppm. By default this plot is not produced for stationary models.

what="statistic"

specifies the empirical and fitted summary statistics, which are plotted using plot.fv. This is only meaningful if the model has been fitted using the Method of Minimum Contrast, and it is turned off otherwise.

Value

Null.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

dppm, plot.ppm, plot.fv.

Examples

  fit <- dppm(swedishpines ~ x + y, dppGauss, method="c")
  plot(fit)

[Package spatstat.model version 3.3-4 Index]