plot.influence.ppm {spatstat.model}	R Documentation

Plot Influence Measure

Description

Plots an influence measure that has been computed by influence.ppm.

Usage

 ## S3 method for class 'influence.ppm'
plot(x, ..., multiplot=TRUE)

Arguments

x
Influence measure (object of class "influence.ppm") computed by influence.ppm.

...
Arguments passed to plot.ppp to control the plotting.

multiplot
Logical value indicating whether it is permissible to plot more than one panel. This happens if the original point process model is multitype.

Details

This is the plot method for objects of class "influence.ppm". These objects are computed by the command influence.ppm.

For a point process model fitted by maximum likelihood or maximum pseudolikelihood (the default), influence values are associated with the data points. The display shows circles centred at the data points with radii proportional to the influence values. If the original data were a multitype point pattern, then if multiplot=TRUE (the default), there is one such display for each possible type of point, while if multiplot=FALSE there is a single plot combining all data points regardless of type.

For a model fitted by logistic composite likelihood (method="logi" in ppm) influence values are associated with the data points and also with the dummy points used to fit the model. The display consist of two panels, for the data points and dummy points respectively, showing circles with radii proportional to the influence values. If the original data were a multitype point pattern, then if multiplot=TRUE (the default), there is one pair of panels for each possible type of point, while if multiplot=FALSE there is a single plot combining all data and dummy points regardless of type.

Use the argument clipwin to restrict the plot to a subset of the full data.

Value

None.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

References

Baddeley, A. and Chang, Y.M. and Song, Y. (2013) Leverage and influence diagnostics for spatial point process models. Scandinavian Journal of Statistics 40, 86–104.

See Also

influence.ppm

Examples

   X <- rpoispp(function(x,y) { exp(3+3*x) })
   fit <- ppm(X, ~x+y)
   plot(influence(fit))

[Package spatstat.model version 3.3-4 Index]