plot.kppm {spatstat.model}	R Documentation

Plot a fitted cluster point process

Description

Plots a fitted cluster point process model, displaying the fitted intensity and the fitted K-function.

Usage

  ## S3 method for class 'kppm'
plot(x, ...,
                      what=c("intensity", "statistic", "cluster"),
                      pause=interactive(),
                      xname)

Arguments

x
Fitted cluster point process model. An object of class "kppm".

...
Arguments passed to plot.ppm and plot.fv to control the plot.

what
Character vector determining what will be plotted.

pause
Logical value specifying whether to pause between plots.

xname
Optional. Character string. The name of the object x for use in the title of the plot.

Details

This is a method for the generic function plot for the class "kppm" of fitted cluster point process models.

The argument x should be a cluster point process model (object of class "kppm") obtained using the function kppm.

The choice of plots (and the order in which they are displayed) is controlled by the argument what. The options (partially matched) are "intensity", "statistic" and "cluster".

This command is capable of producing three different plots:

what="intensity"

specifies the fitted intensity of the model, which is plotted using plot.ppm. By default this plot is not produced for stationary models.

what="statistic"

specifies the empirical and fitted summary statistics, which are plotted using plot.fv. This is only meaningful if the model has been fitted using the Method of Minimum Contrast, and it is turned off otherwise.

what="cluster"

specifies a fitted cluster, which is computed by clusterfield and plotted by plot.im. It is only meaningful for Poisson cluster (incl. Neyman-Scott) processes, and it is turned off for log-Gaussian Cox processes (LGCP). If the model is stationary (and non-LGCP) this option is turned on by default and shows a fitted cluster positioned at the centroid of the observation window. For non-stationary (and non-LGCP) models this option is only invoked if explicitly told so, and in that case an additional argument locations (see clusterfield) must be given to specify where to position the parent point(s) .

Alternatively what="all" selects all available options.

Value

Null.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

kppm, plot.ppm, plot.fv.

Examples

  fit <- kppm(redwood~1, "Thomas")
  plot(fit)

[Package spatstat.model version 3.3-4 Index]