plot.palmdiag {spatstat.model}	R Documentation

Plot the Palm Intensity Diagnostic

Description

Plot the Palm intensity diagnostic for a fitted cluster process or Cox process model.

Usage

## S3 method for class 'palmdiag'
plot(x, ...,
          style = c("intervals", "dots", "bands"),
          args.dots = list(pch = 16), args.intervals = list(),
          xlim=NULL, main)

Arguments

x
Object of class "palmdiag" produced by palmdiagnose.

...
Additional arguments passed to plot.fv when the fitted curve is plotted.

style
Character string specifying the style of plot for the nonparametric estimates. See Details.

args.dots
Arguments passed to points when style="dots".

args.intervals
Arguments passed to segments when style="intervals".

xlim
Optional range of distances plotted along the horizontal axis. A numeric vector of length 2.

main
Optional main title for plot.

Details

This function plots the diagnostic proposed by Tanaka, Ogata and Stoyan (2008, Section 2.3) for assessing goodness-of-fit of a Neyman-Scott cluster process model to a point pattern dataset. The diagnostic is computed by the function palmdiagnose.

First the Palm intensity of the fitted model is plotted as a function of interpoint distance r using plot.fv. Then the nonparametric estimates of the Palm intensity are plotted on the same graph as follows:

if style="dots", the nonparametric estimate for each band of distances is plotted as a dot, with horizontal coordinate at the middle of the band. This is the style proposed by Tanaka et al (2008).

if style="intervals" (the default), each nonparametric estimate is plotted as a dot, and a 95% confidence interval is plotted as a vertical line segment, centred on the dot. The confidence interval is based on the Poisson approximation.

if style="bands", the nonparametric estimates are drawn as a continuous curve which is flat on each band of distances. The 95% confidence intervals are drawn as grey shading.

Value

Null.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Tanaka, U., Ogata, Y. and Stoyan, D. (2008) Parameter estimation and model selection for Neyman-Scott Point Processes. Biometrical Journal 50, 1, 43–57.

See Also

palmdiagnose

Examples

   fit <- kppm(redwood)
   R <- palmdiagnose(fit)
   plot(R, style="d")
   plot(R)
   plot(R, style="b")

[Package spatstat.model version 3.3-4 Index]