ppm.object {spatstat.model}	R Documentation

Class of Fitted Point Process Models

Description

A class ppm to represent a fitted stochastic model for a point process. The output of ppm.

Details

An object of class ppm represents a stochastic point process model that has been fitted to a point pattern dataset. Typically it is the output of the model fitter, ppm.

The class ppm has methods for the following standard generic functions:

generic 	 method 	 description

print 	 print.ppm 	 print details

plot 	 plot.ppm 	 plot fitted model

predict 	 predict.ppm 	 fitted intensity and conditional intensity

fitted 	 fitted.ppm 	 fitted intensity

coef 	 coef.ppm 	 fitted coefficients of model

anova 	 anova.ppm 	 Analysis of Deviance

formula 	 formula.ppm 	 Extract model formula

terms 	 terms.ppm 	 Terms in the model formula

labels 	 labels.ppm 	 Names of estimable terms in the model formula

residuals 	 residuals.ppm 	 Point process residuals

simulate 	 simulate.ppm 	 Simulate the fitted model

update 	 update.ppm 	 Change or refit the model

vcov 	 vcov.ppm 	 Variance/covariance matrix of parameter estimates

model.frame 	 model.frame.ppm 	 Model frame

model.matrix 	 model.matrix.ppm 	 Design matrix

logLik 	 logLik.ppm 	 log pseudo likelihood

extractAIC 	 extractAIC.ppm 	 pseudolikelihood counterpart of AIC

nobs 	 nobs.ppm 	 number of observations

Objects of class ppm can also be handled by the following standard functions, without requiring a special method:

name 	 description

confint 	 Confidence intervals for parameters

step 	 Stepwise model selection

drop1 	 One-step model improvement

add1 	 One-step model improvement

The class ppm also has methods for the following generic functions defined in the spatstat package:

generic 	 method 	 description

as.interact 	 as.interact.ppm 	 Interpoint interaction structure

as.owin 	 as.owin.ppm 	 Observation window of data

berman.test 	 berman.test.ppm 	 Berman's test

envelope 	 envelope.ppm 	 Simulation envelopes

fitin 	 fitin.ppm 	 Fitted interaction

is.marked 	 is.marked.ppm 	 Determine whether the model is marked

is.multitype 	 is.multitype.ppm 	 Determine whether the model is multitype

is.poisson 	 is.poisson.ppm 	 Determine whether the model is Poisson

is.stationary 	 is.stationary.ppm 	 Determine whether the model is stationary

cdf.test 	 cdf.test.ppm 	 Spatial distribution test

quadrat.test 	 quadrat.test.ppm 	 Quadrat counting test

reach 	 reach.ppm 	 Interaction range of model

rmhmodel 	 rmhmodel.ppm 	 Model in a form that can be simulated

rmh 	 rmh.ppm 	 Perform simulation

unitname 	 unitname.ppm 	 Name of unit of length

Information about the data (to which the model was fitted) can be extracted using data.ppm, dummy.ppm and quad.ppm.

Internal format

If you really need to get at the internals, a ppm object contains at least the following entries:

coef 	 the fitted regular parameters (as returned by glm)

trend 	 the trend formula or NULL

interaction 	 the point process interaction family (an object of class "interact") or NULL

Q 	 the quadrature scheme used

maxlogpl 	 the maximised value of log pseudolikelihood

correction 	 name of edge correction method used

See ppm for explanation of these concepts. The irregular parameters (e.g. the interaction radius of the Strauss process) are encoded in the interaction entry. However see the Warnings.

Warnings

The internal representation of ppm objects may change slightly between releases of the spatstat package.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net.

See Also

ppm, coef.ppm, fitted.ppm, print.ppm, predict.ppm, plot.ppm.

Examples

  fit <- ppm(cells ~ x, Strauss(0.1), correction="periodic")
  fit
  coef(fit)
  
    pred <- predict(fit)
  
  pred <- predict(fit, ngrid=20, type="trend")
  if(interactive()) {
  plot(fit)
  }  

[Package spatstat.model version 3.3-4 Index]