ppmInfluence {spatstat.model}	R Documentation

Leverage and Influence Measures for Spatial Point Process Model

Description

Calculates all the leverage and influence measures described in influence.ppm, leverage.ppm and dfbetas.ppm.

Usage

   ppmInfluence(fit,
                what = c("leverage", "influence", "dfbetas"),
                ...,
                iScore = NULL, iHessian = NULL, iArgs = NULL,
                drop = FALSE,
                fitname = NULL)

Arguments

fit
A fitted point process model of class "ppm".

what
Character vector specifying which quantities are to be calculated. Default is to calculate all quantities.

...
Ignored.

iScore,iHessian
Components of the score vector and Hessian matrix for the irregular parameters, if required. See Details.

iArgs
List of extra arguments for the functions iScore, iHessian if required.

drop
Logical. Whether to include (drop=FALSE) or exclude (drop=TRUE) contributions from quadrature points that were not used to fit the model.

fitname
Optional character string name for the fitted model fit.

Details

This function calculates all the leverage and influence measures described in influence.ppm, leverage.ppm and dfbetas.ppm.

When analysing large datasets, the user can call ppmInfluence to perform the calculations efficiently, then extract the leverage and influence values as desired. For example the leverage can be extracted either as result$leverage or leverage(result).

If the point process model trend has irregular parameters that were fitted (using ippm) then the influence calculation requires the first and second derivatives of the log trend with respect to the irregular parameters. The argument iScore should be a list, with one entry for each irregular parameter, of R functions that compute the partial derivatives of the log trend (i.e. log intensity or log conditional intensity) with respect to each irregular parameter. The argument iHessian should be a list, with p^2 entries where p is the number of irregular parameters, of R functions that compute the second order partial derivatives of the log trend with respect to each pair of irregular parameters.

Value

A list containing the leverage and influence measures specified by what. The result also belongs to the class "ppmInfluence".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

leverage.ppm, influence.ppm, dfbetas.ppm

Examples

   X <- rpoispp(function(x,y) { exp(3+3*x) })
   fit <- ppm(X ~ x+y)
   fI <- ppmInfluence(fit)

   fitlev <- fI$leverage
   fitlev <- leverage(fI)

   fitinf <- fI$influence
   fitinf <- influence(fI)

   fitdfb <- fI$dfbetas
   fitdfb <- dfbetas(fI) 

[Package spatstat.model version 3.3-4 Index]