predict.dppm {spatstat.model}	R Documentation

Prediction from a Fitted Determinantal Point Process Model

Description

Given a fitted determinantal point process model, these functions compute the fitted intensity.

Usage

  ## S3 method for class 'dppm'
fitted(object, ...)

  ## S3 method for class 'dppm'
predict(object, ...)

Arguments

object
Fitted determinantal point process model. An object of class "dppm".

...
Arguments passed to fitted.ppm or predict.ppm respectively.

Details

These functions are methods for the generic functions fitted and predict. The argument object should be a determinantal point process model (object of class "dppm") obtained using the function dppm.

The intensity of the fitted model is computed, using fitted.ppm or predict.ppm respectively.

Value

The value of fitted.dppm is a numeric vector giving the fitted values at the quadrature points.

The value of predict.dppm is usually a pixel image (object of class "im"), but see predict.ppm for details.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

dppm, plot.dppm, fitted.ppm, predict.ppm

Examples

  if(interactive()) {
    fit <- dppm(swedishpines ~ x + y, dppGauss())
  } else {
    fit <- dppm(redwood ~ x, dppGauss())
  }
  predict(fit)

[Package spatstat.model version 3.3-4 Index]