predict.kppm {spatstat.model}	R Documentation

Prediction from a Fitted Cluster Point Process Model

Description

Given a fitted cluster point process model, these functions compute the fitted intensity.

Usage

  ## S3 method for class 'kppm'
fitted(object, ...)

  ## S3 method for class 'kppm'
predict(object, ...)

Arguments

object
Fitted cluster point process model. An object of class "kppm".

...
Arguments passed to fitted.ppm or predict.ppm respectively.

Details

These functions are methods for the generic functions fitted and predict. The argument object should be a cluster point process model (object of class "kppm") obtained using the function kppm.

The intensity of the fitted model is computed, using fitted.ppm or predict.ppm respectively.

Value

The value of fitted.kppm is a numeric vector giving the fitted values at the quadrature points.

The value of predict.kppm is usually a pixel image (object of class "im"), but see predict.ppm for details.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

kppm, plot.kppm, vcov.kppm, fitted.ppm, predict.ppm

Examples

  fit <- kppm(redwood ~ x, "Thomas")
  predict(fit)

[Package spatstat.model version 3.3-4 Index]