predict.mppm {spatstat.model}	R Documentation

Prediction for Fitted Multiple Point Process Model

Description

Given a fitted multiple point process model obtained by mppm, evaluate the spatial trend and/or the conditional intensity of the model. By default, predictions are evaluated over a grid of locations, yielding pixel images of the trend and conditional intensity. Alternatively predictions may be evaluated at specified locations with specified values of the covariates.

Usage

## S3 method for class 'mppm'
predict(object, ..., newdata = NULL, type = c("trend", "cif"),
             ngrid = 40, locations=NULL, verbose=FALSE)

Arguments

object
The fitted model. An object of class "mppm" obtained from mppm.

...
Ignored.

newdata
Optional. New values of the covariates, for which the predictions should be computed. See Details.

type
Type of predicted values required. A character string or vector of character strings. Options are "trend" for the spatial trend (first-order term) and "cif" or "lambda" for the conditional intensity. Alternatively type="all" selects all options.

ngrid
Dimensions of the grid of spatial locations at which prediction will be performed (if locations=NULL). An integer or a pair of integers.

locations
Optional. The locations at which predictions should be performed. A list of point patterns, with one entry for each row of newdata.

verbose
Logical flag indicating whether to print progress reports.

Details

This function computes the spatial trend and the conditional intensity of a spatial point process model that has been fitted to several spatial point patterns. See Chapter 16 of Baddeley, Rubak and Turner (2015) for explanation and examples.

Note that by “spatial trend” we mean the (exponentiated) first order potential and not the intensity of the process. [For example if we fit the stationary Strauss process with parameters beta and gamma, then the spatial trend is constant and equal to beta.] The conditional intensity lambda(u,X) of the fitted model is evaluated at each required spatial location u, with respect to the response point pattern X.

If newdata=NULL, predictions are computed for the original values of the covariates, to which the model was fitted. Otherwise newdata should be a hyperframe (see hyperframe) containing columns of covariates as required by the model. If type includes "cif", then newdata must also include a column of spatial point pattern responses, in order to compute the conditional intensity.

If locations=NULL, then predictions are performed at an ngrid by ngrid grid of locations in the window for each response point pattern. The result will be a hyperframe containing a column of images of the trend (if selected) and a column of images of the conditional intensity (if selected). The result can be plotted.

If locations is given, then it should be a list of point patterns (objects of class "ppp"). Predictions are performed at these points, and the results are returned as mark values attached to the locations. The result is a hyperframe containing columns called trend and/or cif. The column called trend contains marked point patterns in which the point locations are the locations and the mark value is the predicted trend. The column called cif contains marked point patterns in which the point locations are the locations and the mark value is the predicted conditional intensity.

Value

A hyperframe with columns named trend and/or cif.

If locations=NULL, the entries of the hyperframe are pixel images.

If locations is not null, the entries are marked point patterns constructed by attaching the predicted values to the locations point patterns.

Models that depend on row number

The point process model that is described by an mppm object may be a different point process for each row of the original hyperframe of data. This occurs if the model formula includes the variable id (representing row number) or if the model has a different interpoint interaction on each row.

If the point process model is different on each row of the original data, then either

newdata is missing. Predictions are computed for each row of the original data using the point process model that applies on each row.

newdata must have the same number of rows as the original data. Each row of newdata is assumed to be a replacement for the corresponding row of the original data. The prediction for row i of newdata will be computed for the point process model that applies to row i of the original data.

newdata must include a column called id specifying the row number, and therefore identifying which of the point process models should apply. The predictions for row i of newdata will be computed for the point process model that applies to row k of the original data, where k = newdata$id[i].

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Ida-Maria Sintorn and Leanne Bischoff. Implemented by Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

References

Baddeley, A. and Turner, R. Practical maximum pseudolikelihood for spatial point patterns. Australian and New Zealand Journal of Statistics 42 (2000) 283–322.

Baddeley, A., Bischof, L., Sintorn, I.-M., Haggarty, S., Bell, M. and Turner, R. Analysis of a designed experiment where the response is a spatial point pattern. In preparation.

Baddeley, A., Rubak, E. and Turner, R. (2015) Spatial Point Patterns: Methodology and Applications with R. Chapman and Hall/CRC Press.

See Also

mppm, fitted.mppm, hyperframe

Examples

  h <- hyperframe(Bugs=waterstriders)
  fit <- mppm(Bugs ~ x, data=h, interaction=Strauss(7))
  # prediction on a grid
  p <- predict(fit)
  plot(p$trend)
  # prediction at specified locations
  loc <- with(h, runifpoint(20, Window(Bugs)))
  p2 <- predict(fit, locations=loc)
  plot(p2$trend)

[Package spatstat.model version 3.3-4 Index]