predict.slrm {spatstat.model}	R Documentation

Predicted or Fitted Values from Spatial Logistic Regression

Description

Given a fitted Spatial Logistic Regression model, this function computes the fitted probabilities for each pixel, or the fitted point process intensity, or the values of the linear predictor in each pixel.

Usage

 ## S3 method for class 'slrm'
predict(object, ..., type = "intensity",
                        newdata=NULL, window=NULL)

Arguments

object
a fitted spatial logistic regression model. An object of class "slrm".

...
Optional arguments passed to pixellate determining the pixel resolution for the discretisation of the point pattern.

type
Character string (partially) matching one of "probabilities", "intensity" or "link".

newdata
Optional. List containing new covariate values for the prediction. See Details.

window
Optional. New window in which to predict. An object of class "owin".

Details

This is a method for predict for spatial logistic regression models (objects of class "slrm", usually obtained from the function slrm).

The argument type determines which quantity is computed. If type="intensity"), the value of the point process intensity is computed at each pixel. If type="probabilities") the probability of the presence of a random point in each pixel is computed. If type="link", the value of the linear predictor is computed at each pixel.

If newdata = NULL (the default), the algorithm computes fitted values of the model (based on the data that was originally used to fit the model object).

If newdata is given, the algorithm computes predicted values of the model, using the new values of the covariates provided by newdata. The argument newdata should be a list; names of entries in the list should correspond to variables appearing in the model formula of the object. Each list entry may be a pixel image or a single numeric value.

Value

A pixel image (object of class "im") containing the predicted values for each pixel.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Rolf Turner rolfturner@posteo.net

See Also

slrm

Examples

  X <- rpoispp(42)
  fit <- slrm(X ~ x+y)
  plot(predict(fit))

  X <- copper$SouthPoints
  Y <- copper$SouthLines
  Z <- distmap(Y)
  fitc <- slrm(X ~ Z)
  pc <- predict(fitc)

  Znew <- distmap(copper$Lines)[copper$SouthWindow]
  pcnew <- predict(fitc, newdata=list(Z=Znew))

[Package spatstat.model version 3.3-4 Index]