print.ppm {spatstat.model}	R Documentation

Print a Fitted Point Process Model

Description

Default print method for a fitted point process model.

Usage

 ## S3 method for class 'ppm'
print(x,...,
     what=c("all", "model", "trend", "interaction", "se", "errors"))

Arguments

x
A fitted point process model, typically obtained from the model-fittingg algorithm ppm. An object of class "ppm".

what
Character vector (partially-matched) indicating what information should be printed.

...
Ignored.

Details

This is the print method for the class "ppm". It prints information about the fitted model in a sensible format.

The argument what makes it possible to print only some of the information.

If what is missing, then by default, standard errors for the estimated coefficients of the model will be printed only if the model is a Poisson point process. To print the standard errors for a non-Poisson model, call print.ppm with the argument what given explicitly, or reset the default rule by typing spatstat.options(print.ppm.SE="always").

Value

none.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

See Also

ppm.object for details of the class "ppm".

ppm for generating these objects.

plot.ppm, predict.ppm

Examples

 
   m <- ppm(cells ~1, Strauss(0.05))
   m
  

[Package spatstat.model version 3.3-4 Index]