prune.rppm {spatstat.model}	R Documentation

Prune a Recursively Partitioned Point Process Model

Description

Given a model which has been fitted to point pattern data by recursive partitioning, apply pruning to reduce the complexity of the partition tree.

Usage

## S3 method for class 'rppm'
prune(tree, ...) 

Arguments

tree
Fitted point process model of class "rppm" produced by the function rppm.

...
Arguments passed to prune.rpart to control the pruning procedure.

Details

This is a method for the generic function prune for the class "rppm". An object of this class is a point process model, fitted to point pattern data by recursive partitioning, by the function rppm.

The recursive partition tree will be pruned using prune.rpart. The result is another object of class "rppm".

Value

Object of class "rppm".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

rppm, plot.rppm, predict.rppm.

Examples

  # Murchison gold data
  mur <- solapply(murchison, rescale, s=1000, unitname="km")
  mur$dfault <- distfun(mur$faults)
  fit <- rppm(gold ~ dfault + greenstone, data=mur)
  fit
  prune(fit, cp=0.1)

[Package spatstat.model version 3.3-4 Index]