pseudoR2 {spatstat.model}	R Documentation

Calculate Pseudo-R-Squared for Point Process Model

Description

Given a fitted point process model, calculate the pseudo-R-squared value, which measures the fraction of variation in the data that is explained by the model.

Usage

  pseudoR2(object, ...)

  ## S3 method for class 'ppm'
pseudoR2(object, ..., keepoffset=TRUE)

  ## S3 method for class 'slrm'
pseudoR2(object, ..., keepoffset=TRUE)

Arguments

object
Fitted point process model. An object of class "ppm" or "slrm".

keepoffset
Logical value indicating whether to retain offset terms in the model when computing the deviance difference. See Details.

...
Additional arguments passed to deviance.ppm or deviance.slrm.

Details

The function pseudoR2 is generic, with methods for fitted point process models of class "ppm" and "slrm".

This function computes McFadden's pseudo-Rsquared

R^2 = 1 - D/D0

where D is the deviance of the fitted model object, and D0 is the deviance of the null model. Deviance is defined as twice the negative log-likelihood or log-pseudolikelihood.

The null model is usually obtained by re-fitting the model using the trend formula ~1. However if the original model formula included offset terms, and if keepoffset=TRUE (the default), then the null model formula consists of these offset terms. This ensures that the pseudoR2 value is non-negative.

Value

A single numeric value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

deviance.ppm, deviance.slrm.

Examples

  fit <- ppm(swedishpines ~ x+y)
  pseudoR2(fit)

  xcoord <- as.im(function(x,y) x, Window(swedishpines))
  fut <- ppm(swedishpines ~ offset(xcoord/200) + y)
  pseudoR2(fut)

[Package spatstat.model version 3.3-4 Index]