psib {spatstat.model}	R Documentation

Sibling Probability of Cluster Point Process

Description

Computes the sibling probability of a cluster point process model.

Usage

  psib(object)

  ## S3 method for class 'kppm'
psib(object)

Arguments

object
Fitted cluster point process model (object of class "kppm").

Details

In a Poisson cluster process, two points are called siblings if they belong to the same cluster, that is, if they had the same parent point. If two points of the process are separated by a distance r, the probability that they are siblings is p(r) = 1 - 1/g(r) where g is the pair correlation function of the process.

The value p(0) = 1 - 1/g(0) is the probability that, if two points of the process are situated very close to each other, they came from the same cluster. This probability is an index of the strength of clustering, with high values suggesting strong clustering.

This concept was proposed in Baddeley, Rubak and Turner (2015, page 479) and Baddeley (2017). It was shown in Baddeley et al (2022) that the sibling probability is directly related to the strength of clustering.

Value

A single number.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Baddeley, A. (2017) Local composite likelihood for spatial point processes. Spatial Statistics 22, 261–295.

Baddeley, A., Rubak, E. and Turner, R. (2015) Spatial Point Patterns: Methodology and Applications with R. Chapman and Hall/CRC Press.

Baddeley, A., Davies, T.M., Hazelton, M.L., Rakshit, S. and Turner, R. (2022) Fundamental problems in fitting spatial cluster process models. Spatial Statistics 52, 100709. DOI: 10.1016/j.spasta.2022.100709

See Also

kppm, panysib

Examples

  fit <- kppm(redwood ~1, "Thomas")
  psib(fit)

[Package spatstat.model version 3.3-4 Index]