rdpp {spatstat.model}	R Documentation

Simulation of a Determinantal Point Process

Description

Generates simulated realisations from a determinantal point process.

Usage

  rdpp(eig, index, basis = "fourierbasis",
       window = boxx(rep(list(0:1), ncol(index))),
       reject_max = 10000, progress = 0, debug = FALSE, ...)

Arguments

eig
vector of values between 0 and 1 specifying the non-zero eigenvalues for the process.

index
data.frame or matrix (or something acceptable to as.matrix) specifying indices of the basis functions.

basis
character string giving the name of the basis.

window
window (of class "owin", "box3" or "boxx") giving the domain of the point process.

reject_max
integer giving the maximal number of trials for rejection sampling.

progress
integer giving the interval for making a progress report. The value zero turns reporting off.

debug
logical value indicating whether debug informationb should be outputted.

...
Ignored.

Value

A point pattern (object of class "ppp").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

Examples

index <- expand.grid(-2:2,-2:2)
eig <- exp(-rowSums(index^2))
X <- rdpp(eig, index)
X
## To simulate a det. projection p. p. with the given indices set eig=1:
XX <- rdpp(1, index)
XX

[Package spatstat.model version 3.3-4 Index]