reach.kppm {spatstat.model}	R Documentation

Range of Interaction for a Cox or Cluster Point Process Model

Description

Returns the range of interaction for a Cox or cluster point process model.

Usage

  ## S3 method for class 'kppm'
reach(x, ..., epsilon)

Arguments

x
Fitted point process model of class "kppm".

epsilon
Optional numerical value. Differences smaller than epsilon are treated as zero.

...
Additional arguments passed to the range function of the given model.

Details

The range of interaction for a fitted point process model of class "kppm" may defined as the smallest number R such that g(r)=1 for all r>=R, where g is the pair correlation function.

For many models the range is infinite, but one may instead use a value where the pair correlation function is sufficiently close to 1. The argument epsilon specifies the tolerance; there is a sensible default.

Value

Numeric

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

Examples

  fit <- kppm(redwood ~ 1)
  reach(fit)

[Package spatstat.model version 3.3-4 Index]