residualMeasure {spatstat.model}	R Documentation

Residual Measure for an Observed Point Pattern and a Fitted Intensity

Description

Given a point pattern and an estimate of its intensity function obtained in any fashion, compute the residual measure.

Usage

residualMeasure(Q, lambda,
               type = c("raw", "inverse", "Pearson", "pearson"),
               ...)

Arguments

Q
A point pattern (object of class "ppp") or quadrature scheme (object of class "quad").

lambda
Predicted intensity. An image (object of class "im") or a list of images.

type
Character string (partially matched) specifying the type of residuals.

...
Arguments passed to quadscheme if Q is a point pattern.

Details

This command constructs the residual measure for the model in which Q is the observed point pattern or quadrature scheme, and lambda is the estimated intensity function obtained in some fashion.

Value

A measure (object of class "msr").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Baddeley, A., Turner, R., Moller, J. and Hazelton, M. (2005) Residual analysis for spatial point processes. Journal of the Royal Statistical Society, Series B 67, 617–666.

Baddeley, A., Moller, J. and Pakes, A.G. (2008) Properties of residuals for spatial point processes. Annals of the Institute of Statistical Mathematics 60, 627–649.

See Also

residuals.ppm

Examples

  ## nonparametric regression estimate of intensity
  ## as a function of terrain elevation
  f <- rhohat(bei, bei.extra$elev)
  ## predicted intensity as a function of location
  lam <- predict(f)
  ## residuals
  res <- residualMeasure(bei, lam)
  res
  plot(res)

[Package spatstat.model version 3.3-4 Index]