residuals.kppm {spatstat.model}	R Documentation

Residuals for Fitted Cox or Cluster Point Process Model

Description

Given a Cox or cluster point process model fitted to a point pattern, compute residuals.

Usage

  ## S3 method for class 'kppm'
residuals(object, ...)

Arguments

object
The fitted point process model (an object of class "kppm") for which residuals should be calculated.

...
Arguments passed to residuals.ppm.

Details

This function extracts the intensity component of the model using as.ppm and then applies residuals.ppm to compute the residuals.

Use plot.msr to plot the residuals directly.

Value

An object of class "msr" representing a signed measure or vector-valued measure (see msr). This object can be plotted.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

See Also

msr, kppm

Examples

   fit <- kppm(redwood ~ x, "Thomas")
   rr <- residuals(fit)

[Package spatstat.model version 3.3-4 Index]