residuals.rppm {spatstat.model}	R Documentation

Residuals for Recursively Partitioned Point Process Model

Description

Given a point process model that was fitted to a point pattern by recursive partitioning (regression tree) methods, compute the residual measure.

Usage

  ## S3 method for class 'rppm'
residuals(object,
                           type=c("raw", "inverse", "Pearson"),
                           ...)

Arguments

object
The fitted point process model (an object of class "ppm") for which residuals should be calculated.

type
String (partially matched) indicating the type of residuals to be calculated.

...
Ignored.

Details

This function computes the residual measure for a point process model that was fitted to point pattern data by recursive partitioning of the covariates.

The argument object must be a fitted model object of class "rppm"). Such objects are created by the fitting algorithm rppm.

The type of residual is chosen by the argument type.

Value

A measure (object of class "msr").

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Baddeley, A., Turner, R., Moller, J. and Hazelton, M. (2005) Residual analysis for spatial point processes. Journal of the Royal Statistical Society, Series B 67, 617–666.

Baddeley, A., Moller, J. and Pakes, A.G. (2008) Properties of residuals for spatial point processes. Annals of the Institute of Statistical Mathematics 60, 627–649.

See Also

residuals.ppm

Examples

  fit <- rppm(bei ~ elev + grad, data=bei.extra)
  res <- residuals(fit)
  plot(res)

[Package spatstat.model version 3.3-4 Index]