response {spatstat.model}	R Documentation

Extract the Values of the Response from a Fitted Model

Description

Given a fitted model (of any kind) extract the values of the response variable. For a point process model, the observed point pattern is extracted.

Usage

response(object)

## S3 method for class 'lm'
response(object)

## S3 method for class 'glm'
response(object)

## S3 method for class 'ppm'
response(object)

## S3 method for class 'kppm'
response(object)

## S3 method for class 'dppm'
response(object)

## S3 method for class 'slrm'
response(object)

## S3 method for class 'rppm'
response(object)

## S3 method for class 'mppm'
response(object)

Arguments

object
A fitted model (object of class "lm", "glm", "ppm", "kppm", "dppm", "slrm", "rppm", or "mppm" or some other class).

Details

For fitted linear models of class "lm" and fitted generalized linear models of class "glm", the numerical values of the response variable are extracted if they are available, and otherwise NULL is returned.

For fitted point process models of class "ppm", "kppm", "dppm", "slrm", "lppm" or "rppm", the original data point pattern is extracted.

For a fitted point process model of class "mppm", the list of original data point patterns is extracted.

Value

For response.lm and response.glm, a numeric vector, or NULL.

For response.ppm, response.kppm, response.dppm, response.slrm and response.rppm, a two-dimensional spatial point pattern (class "ppp").

For response.mppm, a list of two-dimensional spatial point patterns (objects of class "ppp"). The list also belongs to classes "solist" and "ppplist".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

Examples

  fit <- ppm(cells ~ x)
  response(fit)  

[Package spatstat.model version 3.3-4 Index]