subfits {spatstat.model}	R Documentation

Extract List of Individual Point Process Models

Description

Takes a Gibbs point process model that has been fitted to several point patterns simultaneously, and produces a list of fitted point process models for the individual point patterns.

Usage

   subfits(object, what="models", verbose=FALSE, new.coef=NULL)
   subfits.old(object, what="models", verbose=FALSE, new.coef=NULL)
   subfits.new(object, what="models", verbose=FALSE)

Arguments

object
An object of class "mppm" representing a point process model fitted to several point patterns.

what
What should be returned. Either "models" to return the fitted models, or "interactions" to return the fitted interactions only.

verbose
Logical flag indicating whether to print progress reports.

new.coef
Advanced use only. Numeric vector or matrix of coefficients to replaced the fitted coefficients coef(object).

Details

object is assumed to have been generated by mppm. It represents a point process model that has been fitted to a list of several point patterns, with covariate data.

For each of the individual point pattern datasets, this function derives the corresponding fitted model for that dataset only (i.e. a point process model for the ith point pattern, that is consistent with object).

If what="models", the result is a list of point process models (a list of objects of class "ppm"), one model for each point pattern dataset in the original fit. If what="interactions", the result is a list of fitted interpoint interactions (a list of objects of class "fii").

Two different algorithms are provided, as subfits.old and subfits.new. Currently subfits is the same as the old algorithm subfits.old because the newer algorithm is too memory-hungry.

Value

A list of point process models (a list of objects of class "ppm") or a list of fitted interpoint interactions (a list of objects of class "fii").

Author(s)

Adrian Baddeley, Ida-Maria Sintorn and Leanne Bischoff. Implemented in spatstat by Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

References

Baddeley, A., Rubak, E. and Turner, R. (2015) Spatial Point Patterns: Methodology and Applications with R. Chapman and Hall/CRC Press.

See Also

mppm, ppm

Examples

  H <- hyperframe(Wat=waterstriders)
  fit <- mppm(Wat~x, data=H)
  subfits(fit)

  H$Wat[[3]] <- rthin(H$Wat[[3]], 0.1)
  fit2 <- mppm(Wat~x, data=H, random=~1|id)
  subfits(fit2)

  

[Package spatstat.model version 3.3-4 Index]