summary.dppm {spatstat.model}	R Documentation

Summarizing a Fitted Determinantal Point Process Model

Description

summary method for class "dppm".

Usage

  ## S3 method for class 'dppm'
summary(object, ..., quick=FALSE)

  ## S3 method for class 'summary.dppm'
print(x, ...)

Arguments

object
A fitted determinantal point process model (object of class "dppm").

quick
Logical value controlling the scope of the summary.

...
Arguments passed to summary.ppm or print.summary.ppm controlling the treatment of the trend component of the model.

x
Object of class "summary.dppm" as returned by summary.dppm.

Details

This is a method for the generic summary for the class "dppm". An object of class "dppm" describes a fitted determinantal point process model. See dppm.

summary.dppm extracts information about the type of model that has been fitted, the data to which the model was fitted, and the values of the fitted coefficients.

print.summary.dppm prints this information in a comprehensible format.

In normal usage, print.summary.dppm is invoked implicitly when the user calls summary.dppm without assigning its value to anything. See the examples.

Value

summary.dppm returns an object of class "summary.dppm", while print.summary.dppm returns NULL.

The result of summary.dppm includes at least the following components:

Xname
character string name of the original point pattern data

stationary
logical value indicating whether the model is stationary

trend
Object of class summary.ppm summarising the trend

repul
Repulsiveness index

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

Examples

  jpines <- residualspaper$Fig1
  
 fit <- dppm(jpines ~ 1, dppGauss)
 summary(fit)

[Package spatstat.model version 3.3-4 Index]