summary.kppm {spatstat.model}	R Documentation

Summarizing a Fitted Cox or Cluster Point Process Model

Description

summary method for class "kppm".

Usage

  ## S3 method for class 'kppm'
summary(object, ..., quick=FALSE)

  ## S3 method for class 'summary.kppm'
print(x, ...)

Arguments

object
A fitted Cox or cluster point process model (object of class "kppm").

quick
Logical value controlling the scope of the summary.

...
Arguments passed to summary.ppm or print.summary.ppm controlling the treatment of the trend component of the model.

x
Object of class "summary.kppm" as returned by summary.kppm.

Details

This is a method for the generic summary for the class "kppm". An object of class "kppm" describes a fitted Cox or cluster point process model. See kppm.

summary.kppm extracts information about the type of model that has been fitted, the data to which the model was fitted, and the values of the fitted coefficients.

print.summary.kppm prints this information in a comprehensible format.

In normal usage, print.summary.kppm is invoked implicitly when the user calls summary.kppm without assigning its value to anything. See the examples.

You can also type coef(summary(object)) to extract a table of the fitted coefficients of the point process model object together with standard errors and confidence limits.

Value

summary.kppm returns an object of class "summary.kppm", while print.summary.kppm returns NULL.

The result of summary.kppm includes at least the following components:

Xname
character string name of the original point pattern data

stationary
logical value indicating whether the model is stationary

clusters
the clusters argument to kppm

modelname
character string describing the model

isPCP
TRUE if the model is a Poisson cluster process, FALSE if it is a log-Gaussian Cox process

lambda
Estimated intensity: numeric value, or pixel image

mu
Mean cluster size: numeric value, pixel image, or NULL

clustpar
list of fitted parameters for the cluster model

clustargs
list of fixed parameters for the cluster model, if any

callstring
character string representing the original call to kppm

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

References

Baddeley, A., Davies, T.M., Hazelton, M.L., Rakshit, S. and Turner, R. (2022) Fundamental problems in fitting spatial cluster process models. Spatial Statistics 52, 100709. DOI: 10.1016/j.spasta.2022.100709

Examples

 fit <- kppm(redwood ~ 1, "Thomas")
 summary(fit)
 coef(summary(fit))

[Package spatstat.model version 3.3-4 Index]