traj {spatstat.model}	R Documentation

Extract trajectory of function evaluations

Description

Extract the history of evaluations of the objective function performed when a cluster process model was fitted.

Usage

traj(object)

Arguments

object
Fitted cluster point process model (object of class "kppm") or objective function surface (object of class "objsurf").

Details

Under appropriate circumstances, the fitted model object contains the history of evaluations of the objective function that were performed by the optimisation algorithm. This history is extracted by traj.

The result is a data frame containing the input parameter values for the objective function, and the corresponding value of the objective function, that were considered by the optimisation algorithm. This data frame also belongs to the class "traj" which has methods for plot, print and other purposes.

Value

Either a data frame (belonging to class "traj") or NULL.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

methods.traj

Examples

  fit <- kppm(redwood, trajectory=TRUE)
  h <- traj(fit)

[Package spatstat.model version 3.3-4 Index]