triplet.family {spatstat.model}	R Documentation

Triplet Interaction Family

Description

An object describing the family of all Gibbs point processes with interaction order equal to 3.

Details

Advanced Use Only!

This structure would not normally be touched by the user. It describes the interaction structure of Gibbs point processes which have infinite order of interaction, such as the triplet interaction process Triplets.

Anyway, triplet.family is an object of class "isf" containing a function triplet.family$eval for evaluating the sufficient statistics of a Gibbs point process model taking an exponential family form.

Value

Object of class "isf", see isf.object.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net

References

Baddeley, A. and Turner, R. (2000) Practical maximum pseudolikelihood for spatial point patterns. Australian and New Zealand Journal of Statistics 42, 283–322.

See Also

Triplets to create the triplet interaction process structure.

Other families: pairwise.family, pairsat.family, inforder.family, ord.family.

[Package spatstat.model version 3.3-4 Index]