update.detpointprocfamily {spatstat.model}	R Documentation

Set Parameter Values in a Determinantal Point Process Model

Description

Set parameter values in a determinantal point process model object.

Usage

  ## S3 method for class 'detpointprocfamily'
update(object, ...)

Arguments

object
object of class "detpointprocfamily".

...
arguments of the form tag=value specifying the parameters values to set.

Value

Another object of class "detpointprocfamily".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

[Package spatstat.model version 3.3-4 Index]