update.dppm {spatstat.model}	R Documentation

Update a Fitted Determinantal Point Process Model

Description

update method for class "dppm".

Usage

 ## S3 method for class 'dppm'
update(object, ..., evaluate=TRUE,
                       envir=environment(terms(object)))

Arguments

object
Fitted determinantal point process model. An object of class "dppm", obtained from dppm.

...
Arguments passed to dppm.

evaluate
Logical value indicating whether to return the updated fitted model (evaluate=TRUE, the default) or just the updated call to dppm (evaluate=FALSE).

envir
Environment in which to re-evaluate the call to dppm.

Details

object should be a fitted determinantal point process model, obtained from the model-fitting function dppm. The model will be updated according to the new arguments provided.

If the argument trend is provided, it determines the intensity in the updated model. It should be an R formula (with or without a left hand side). It may include the symbols + or - to specify addition or deletion of terms in the current model formula, as shown in the Examples below. The symbol . refers to the current contents of the formula.

The intensity in the updated model is determined by the argument trend if it is provided, or otherwise by any unnamed argument that is a formula, or otherwise by the formula of the original model, formula(object).

The spatial point pattern data to which the new model is fitted is determined by the left hand side of the updated model formula, if this is present. Otherwise it is determined by the argument X if it is provided, or otherwise by any unnamed argument that is a point pattern or a quadrature scheme.

The model is refitted using dppm.

Value

Another fitted cluster point process model (object of class "dppm".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

dppm, plot.dppm, predict.dppm, simulate.dppm, methods.dppm.

Examples

  fit <- dppm(swedishpines ~ x + y, dppGauss, method="c")
  fitx <- update(fit, ~x)
  fit2 <- update(fit, flipxy(swedishpines))

[Package spatstat.model version 3.3-4 Index]