update.interact {spatstat.model}	R Documentation

Update an Interpoint Interaction

Description

This command updates the object using the arguments given.

Usage

## S3 method for class 'interact'
update(object, ...)

Arguments

object
Interpoint interaction (object of class "interact").

...
Additional or replacement values of parameters of object.

Details

This is a method for the generic function update for the class "interact" of interpoint interactions. It updates the object using the parameters given in the extra arguments ....

The extra arguments must be given in the form name=value and must be recognisable to the interaction object. They override any parameters of the same name in object.

Value

Another object of class "interact", equivalent to object except for changes in parameter values.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

update.ppm

Examples

  Str <- Strauss(r=1)
  Str
  update(Str, r=2)

  M <- MultiStrauss(radii=matrix(1,2,2))
  update(M, types=c("on", "off"))

[Package spatstat.model version 3.3-4 Index]