update.rppm {spatstat.model}	R Documentation

Update a Recursively Partitioned Point Process Model

Description

update method for class "rppm".

Usage

 ## S3 method for class 'rppm'
update(object, ..., envir=environment(terms(object)))

Arguments

object
Fitted recursively partitioned point process model. An object of class "rppm", obtained from rppm.

...
Arguments passed to rppm.

envir
Environment in which to re-evaluate the call to rppm.

Details

object should be a fitted recursively partitioned point process model, obtained from the model-fitting function rppm.

The model will be updated according to the new arguments provided.

Value

Another fitted recursively partitioned point process model (object of class "rppm".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

rppm.

Examples

 fit <- rppm(nztrees ~ x)
 newfit <- update(fit, . ~ x + y)

[Package spatstat.model version 3.3-4 Index]