valid {spatstat.model}	R Documentation

Check Whether Point Process Model is Valid

Description

Determines whether a point process model object corresponds to a valid point process.

Usage

  valid(object, ...)

Arguments

object
Object of some class, describing a point process model.

...
Additional arguments passed to methods.

Details

The function valid is generic, with methods for the classes "ppm" and "dppmodel".

An object representing a point process is called valid if all its parameter values are known (for example, no parameter takes the value NA or NaN) and the parameter values correspond to a well-defined point process (for example, the parameter values satisfy all the constraints that are imposed by mathematical theory.)

See the methods for further details.

Value

A logical value, or NA.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

Rolf Turner rolfturner@posteo.net

and Ege Rubak rubak@math.aau.dk

See Also

valid.ppm, valid.detpointprocfamily

[Package spatstat.model version 3.3-4 Index]