varcount {spatstat.model}	R Documentation

Predicted Variance of the Number of Points

Description

Given a fitted point process model, calculate the predicted variance of the number of points in a nominated set B.

Usage

varcount(model, B=Window(model), ..., dimyx = NULL, relative=FALSE)

Arguments

model
A fitted point process model (object of class "ppm", "kppm" or "dppm").

B
A window (object of class "owin" specifying the region in which the points are counted. Alternatively a pixel image (object of class "im") or a function of spatial coordinates specifying a numerical weight for each random point. The default is the window of the original point pattern data to which the model was fitted.

...
Additional arguments passed to B when it is a function.

dimyx
Spatial resolution for the calculations. Argument passed to as.mask.

relative
Logical value specifying whether to divide the variance by the mean value.

Details

The function varcount calculates the variance of the number of points falling in a specified window B according to the model. It can also calculate the variance of a sum of weights attached to each random point.

If relative=FALSE (the default), the result is the variance. If relative=TRUE, the result is the variance divided by the mean, which is the overdispersion index (equal to 1 if the number of points has a Poisson distribution).

The model should be a fitted point process model (object of class "ppm", "kppm" or "dppm").

If B is a window, varcount calculates the variance of the number of points falling in B, according to the fitted model.

If the model depends on spatial covariates other than the Cartesian coordinates, then B should be a subset of the domain in which these covariates are defined.

If B is a pixel image, varcount calculates the variance of T = sum[i] B(x[i]), the sum of the values of B over all random points falling in the domain of the image.

If the model depends on spatial covariates other than the Cartesian coordinates, then the domain of the pixel image, as.owin(B), should be a subset of the domain in which these covariates are defined.

If B is a function(x,y) or function(x,y,...) then varcount calculates the variance of T = sum[i] B(x[i]), the sum of the values of B over all random points falling inside the window W=as.owin(model), the window in which the original data were observed.

The variance calculation involves the intensity and the pair correlation function of the model. The calculation is exact (up to discretisation error) for models of class "kppm" and "dppm", and for Poisson point process models of class "ppm". For Gibbs point process models of class "ppm" the calculation depends on the Poisson-saddlepoint approximations to the intensity and pair correlation function, which are rough approximations. The approximation is not yet implemented for some Gibbs models.

Value

A single number.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk

See Also

predict.ppm, predict.kppm, predict.dppm

Examples

   fitT <- kppm(redwood ~ 1, "Thomas")
   B <- owin(c(0, 0.5), c(-0.5, 0))
   varcount(fitT, B)

   fitS <- ppm(swedishpines ~ 1, Strauss(9))
   BS <- square(50)
   varcount(fitS, BS)

[Package spatstat.model version 3.3-4 Index]