vcov.kppm {spatstat.model}	R Documentation

Variance-Covariance Matrix for a Fitted Cluster Point Process Model

Description

Returns the variance-covariance matrix of the estimates of the parameters of a fitted cluster point process model.

Usage

   ## S3 method for class 'kppm'
vcov(object, ...,
          what=c("vcov", "corr", "fisher"),
          fast = NULL, rmax = NULL, eps.rmax = 0.01,
          verbose = TRUE)

Arguments

object
A fitted cluster point process model (an object of class "kppm".)

...
Ignored.

what
Character string (partially-matched) that specifies what matrix is returned. Options are "vcov" for the variance-covariance matrix, "corr" for the correlation matrix, and "fisher" for the Fisher information matrix.

fast
Logical specifying whether tapering (using sparse matrices from Matrix) should be used to speed up calculations. Warning: This is expected to underestimate the true asymptotic variances/covariances.

rmax
Optional. The dependence range. Not usually specified by the user. Only used when fast=TRUE.

eps.rmax
Numeric. A small positive number which is used to determine rmax from the tail behaviour of the pair correlation function when fast option (fast=TRUE) is used. Namely rmax is the smallest value of r at which (g(r)-1)/(g(0)-1) falls below eps.rmax. Only used when fast=TRUE. Ignored if rmax is provided.

verbose
Logical value indicating whether to print progress reports during very long calculations.

Details

This function computes the asymptotic variance-covariance matrix of the estimates of the canonical (regression) parameters in the cluster point process model object. It is a method for the generic function vcov.

The result is an n * n matrix where n = length(coef(model)).

To calculate a confidence interval for a regression parameter, use confint as shown in the examples.

Value

A square matrix.

Author(s)

Abdollah Jalilian and Rasmus Waagepetersen. Ported to spatstat by Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Ege Rubak rubak@math.aau.dk.

References

Waagepetersen, R. (2007) Estimating functions for inhomogeneous spatial point processes with incomplete covariate data. Biometrika 95, 351–363.

See Also

kppm, vcov, vcov.ppm

Examples

   fit <- kppm(redwood ~ x + y)
   vcov(fit)
   vcov(fit, what="corr")

   # confidence interval
   confint(fit)
   # cross-check the confidence interval by hand:
   sd <- sqrt(diag(vcov(fit)))
   t(coef(fit) + 1.96 * outer(sd, c(lower=-1, upper=1)))

[Package spatstat.model version 3.3-4 Index]