vcov.slrm {spatstat.model}	R Documentation

Variance-Covariance Matrix for a Fitted Spatial Logistic Regression

Description

Returns the variance-covariance matrix of the estimates of the parameters of a point process model that was fitted by spatial logistic regression.

Usage

  ## S3 method for class 'slrm'
vcov(object, ...,
         what=c("vcov", "corr", "fisher", "Fisher")) 

Arguments

object
A fitted point process model of class "slrm".

...
Ignored.

what
Character string (partially-matched) that specifies what matrix is returned. Options are "vcov" for the variance-covariance matrix, "corr" for the correlation matrix, and "fisher" or "Fisher" for the Fisher information matrix.

Details

This function computes the asymptotic variance-covariance matrix of the estimates of the canonical parameters in the point process model object. It is a method for the generic function vcov.

object should be an object of class "slrm", typically produced by slrm. It represents a Poisson point process model fitted by spatial logistic regression.

The canonical parameters of the fitted model object are the quantities returned by coef.slrm(object). The function vcov calculates the variance-covariance matrix for these parameters.

The argument what provides three options:

what="vcov"

return the variance-covariance matrix of the parameter estimates

what="corr"

return the correlation matrix of the parameter estimates

what="fisher"

return the observed Fisher information matrix.

In all three cases, the result is a square matrix. The rows and columns of the matrix correspond to the canonical parameters given by coef.slrm(object). The row and column names of the matrix are also identical to the names in coef.slrm(object).

Note that standard errors and 95% confidence intervals for the coefficients can also be obtained using confint(object) or coef(summary(object)).

Standard errors for the fitted intensity can be obtained using predict.slrm.

Value

A square matrix.

Error messages

An error message that reports system is computationally singular indicates that the determinant of the Fisher information matrix was either too large or too small for reliable numerical calculation. This can occur because of numerical overflow or collinearity in the covariates.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

and Rolf Turner rolfturner@posteo.net .

References

Baddeley, A., Berman, M., Fisher, N.I., Hardegen, A., Milne, R.K., Schuhmacher, D., Shah, R. and Turner, R. (2010) Spatial logistic regression and change-of-support for spatial Poisson point processes. Electronic Journal of Statistics 4, 1151–1201. DOI: 10.1214/10-EJS581

See Also

vcov for the generic,

slrm for information about fitted models,

predict.slrm for other kinds of calculation about the model,

confint for confidence intervals.

Examples

  X <- rpoispp(42)
  fit <- slrm(X ~ x + y)
  vcov(fit)
  vcov(fit, what="corr")
  vcov(fit, what="f")

[Package spatstat.model version 3.3-4 Index]