zgibbsmodel {spatstat.model}	R Documentation

Gibbs Model

Description

Experimental code. Creates an object representing a Gibbs point process model. Typically used for theoretical calculations about such a model.

Usage

zgibbsmodel(beta = 1, interaction = NULL, icoef = NULL)

Arguments

beta
First order trend term. A numeric value, numeric vector, pixel image, function, or a list of such objects.

interaction
Object of class "interact" specifying the interpoint interaction structure, or NULL representing the Poisson process.

icoef
Numeric vector of coefficients for the interpoint interaction.

Details

Experimental.

Value

Object belonging to the experimental class zgibbsmodel.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

See Also

methods.zgibbsmodel

Examples

  m <- zgibbsmodel(10, Strauss(0.1), -0.5)

[Package spatstat.model version 3.3-4 Index]