Kmodel {spatstat.model}	R Documentation

K Function or Pair Correlation Function of a Point Process Model

Description

Returns the theoretical K function or the pair correlation function of a point process model.

Usage

   Kmodel(model, ...)

   pcfmodel(model, ...)

Arguments

model
A fitted point process model of some kind.

...
Ignored.

Details

For certain types of point process models, it is possible to write down a mathematical expression for the K function or the pair correlation function of the model.

The functions Kmodel and pcfmodel give the theoretical K-function and the theoretical pair correlation function for a point process model that has been fitted to data.

The functions Kmodel and pcfmodel are generic, with methods for the classes "kppm" (cluster processes and Cox processes) and "ppm" (Gibbs processes).

The return value is a function in the R language, which takes one argument r. Evaluation of this function, on a numeric vector r, yields values of the desired K function or pair correlation function at these distance values.

Value

A function in the R language, which takes one argument r.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

Kest or pcf to estimate the K function or pair correlation function nonparametrically from data.

Kmodel.kppm for the method for cluster processes and Cox processes.

Kmodel.ppm for the method for Gibbs processes.

[Package spatstat.model version 3.3-4 Index]