Kmodel.ppm {spatstat.model}	R Documentation

K Function or Pair Correlation Function of Gibbs Point Process model

Description

Returns the theoretical K function or the pair correlation function of a fitted Gibbs point process model.

Usage

   ## S3 method for class 'ppm'
Kmodel(model, ...)

   ## S3 method for class 'ppm'
pcfmodel(model, ...)

Arguments

model
A fitted Poisson or Gibbs point process model (object of class "ppm") typically obtained from the model-fitting algorithm ppm.

...
Ignored.

Details

This function computes an approximation to the K function or the pair correlation function of a Gibbs point process.

The functions Kmodel and pcfmodel are generic. The functions documented here are the methods for the class "ppm".

The approximation is only available for stationary pairwise-interaction models. It uses the second order Poisson-saddlepoint approximation (Baddeley and Nair, 2012b) which is a combination of the Poisson-Boltzmann-Emden and Percus-Yevick approximations.

The return value is a function in the R language, which takes one argument r. Evaluation of this function, on a numeric vector r, yields values of the desired K function or pair correlation function at these distance values.

Value

A function in the R language, which takes one argument r.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Gopalan Nair.

References

Baddeley, A. and Nair, G. (2012a) Fast approximation of the intensity of Gibbs point processes. Electronic Journal of Statistics 6 1155–1169.

Baddeley, A. and Nair, G. (2012b) Approximating the moments of a spatial point process. Stat 1, 1, 18–30. DOI: 10.1002/sta4.5

See Also

Kest or pcf to estimate the K function or pair correlation function nonparametrically from data.

ppm to fit Gibbs models.

Kmodel for the generic functions.

Kmodel.kppm for the method for cluster/Cox processes.

Examples

  fit <- ppm(swedishpines, ~1, Strauss(8))
  p <- pcfmodel(fit)
  K <- Kmodel(fit)
  p(6)
  K(8)
  curve(K(x), from=0, to=15)

[Package spatstat.model version 3.3-4 Index]