repul.dppm {spatstat.model}	R Documentation

Repulsiveness Index of a Determinantal Point Process Model

Description

Computes a measure of the degree of repulsion between points in a determinantal point process model.

Usage

repul(model, ...)

## S3 method for class 'dppm'
repul(model, ...)

Arguments

model
A fitted point process model of determinantal type (object of class "dppm").

...
Ignored.

Details

The repulsiveness index mu of a determinantal point process model was defined by Lavancier, Moller and Rubak (2015) as

mu = lambda * integral[(1-g(x))]

where λ is the intensity of the model and g(x) is the pair correlation function, and the integral is taken over all two-dimensional vectors x.

Values of μ are dimensionless. Larger values of mu indicate stronger repulsion between points.

If the model is stationary, the result is a single number.

If the model is not stationary, the result is a pixel image (obtained by multiplying the spatially-varying intensity by the integral defined above).

Value

A numeric value or a pixel image.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Lavancier, F., Moller, J. and Rubak, E. (2015), Determinantal point process models and statistical inference. Journal of Royal Statistical Society: Series B (Statistical Methodology), 77, 853–877.

See Also

dppm

Examples

  jpines <- residualspaper$Fig1
  
  fit <- dppm(jpines ~ 1, dppGauss)
  repul(fit)

[Package spatstat.model version 3.3-4 Index]