zclustermodel {spatstat.model}	R Documentation

Cluster Point Process Model

Description

Experimental code. Creates an object representing a cluster point process model. Typically used for theoretical calculations about such a model.

Usage

zclustermodel(name = "Thomas", ..., mu, kappa, scale)

Arguments

name
Name of the cluster process. One of "Thomas", "MatClust", "VarGamma" or "Cauchy".

...
Other arguments needed for the model.

mu
Mean cluster size. A single number, or a pixel image.

kappa
Parent intensity. A single number.

scale
Cluster scale parameter of the model.

Details

Experimental.

Value

Object of the experimental class "zclustermodel".

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

See Also

methods.zclustermodel

Examples

  m <- zclustermodel("Thomas", kappa=10, mu=5, scale=0.1)

[Package spatstat.model version 3.3-4 Index]